### 2025-06-09 #8375 @editorialbot
[PRE REVIEW]: Enhancing Next-Generation Sequencing Simulation: Updates to NEAT
**Submitting author:** @keshav-gandhi (<a href="http://orcid.org/0009-0000-1718-1862">Keshav Gandhi</a>)
**Repository:** https://github.com/ncsa/NEAT
**Branch with paper.md** (empty if default branch): paper
**Version:** 4.2
