### 2025-07-10 1b7935b
Stitching output script implementation.
