+++ neat/read_simulator/utils/stitch_outputs.py
@@ -1,0 +1,137 @@
+from __future__ import annotations
+
+import argparse
+import gzip
+import shutil
+import subprocess
+import sys
+from pathlib import Path
+from typing import Iterable, List
+import yaml
+
+__all__ = ["main"]
+
+
+def print_stderr(msg: str, *, exit_: bool = False):
+    """Print to stderr, with an option to exit."""
+    print(msg, file=sys.stderr)
+    if exit_:
+        sys.exit(1)
+
+
+def gather(paths: Iterable[Path], suffixes: tuple[str, ...]) -> List[Path]:
+    """Gather all files in the provided path whose name ends with the given suffix."""
+    out: list[Path] = []
+    for p in paths:
+        if p.is_file() and p.name.endswith(suffixes):
+            out.append(p)
+        elif p.is_dir():
+            for suf in suffixes:
+                out.extend(p.rglob(f"*{suf}"))
+    return sorted(set(out))
+
+
+def is_gzipped(file: Path) -> bool:
+    return file.suffix in {".gz", ".bgz"}
+
+
+def concat(files: List[Path], dest: Path, keep_compression: bool):
+    if not files:
+        return
+    dest.parent.mkdir(parents=True, exist_ok=True)
+
+    if keep_compression:
+        if not dest.suffix.endswith("gz"):
+            dest = dest.with_suffix(dest.suffix + ".gz")
+        with dest.open("wb") as out_f:
+            for f in files:
+                with f.open("rb") as in_f:
+                    shutil.copyfileobj(in_f, out_f)
+    else:
+        # stream out uncompressed lines for speed
+        with dest.open("wb") as out_f:
+            for f in files:
+                opener = gzip.open if is_gzipped(f) else open
+                with opener(f, "rb") as in_f:
+                    shutil.copyfileobj(in_f, out_f)
+
+
+def merge_bam(bams: List[Path], dest: Path, samtools: str):
+    if not bams:
+        return
+    dest.parent.mkdir(parents=True, exist_ok=True)
+    unsorted = dest.with_suffix(".unsorted.bam")
+    cmd_merge = [samtools, "merge", "--no-PG", "-f", str(unsorted), *map(str, bams)]
+    subprocess.check_call(cmd_merge)
+    subprocess.check_call([samtools, "sort", "-o", str(dest), str(unsorted)])
+    subprocess.check_call([samtools, "index", str(dest)])
+    unsorted.unlink(missing_ok=True)
+
+
+def merge_vcf(vcfs: List[Path], dest: Path):
+    if not vcfs:
+        return
+    dest.parent.mkdir(parents=True, exist_ok=True)
+    first, *rest = vcfs
+    shutil.copy(first, dest)
+    with dest.open("ab") as out_f:
+        for vcf in rest:
+            with vcf.open("rb") as fh:
+                for line in fh:
+                    if not line.startswith(b"#"):
+                        out_f.write(line)
+
+
+def parse_args(argv: list[str] | None = None):
+    p = argparse.ArgumentParser(description="Stitch NEAT split‑run outputs into one dataset.")
+    p.add_argument("-i", "--inputs", nargs="+", type=Path, required=True,
+                   help="Input split‑run directories or file prefixes")
+    p.add_argument("-o", "--output-prefix", type=Path, required=True,
+                   help="Prefix for stitched outputs (no extension)")
+    p.add_argument("-c", "--config", type=Path, required=True,
+                   help="Original NEAT YAML config to detect paired‑ended, etc.")
[…]
