### 2025-08-04 f9931f0
Added option to remove temporary files.
