### 2025-08-04 b1fa382
Option to not split if already done.
