### 2025-08-12 e0e79a6
Double-checking exactness between constant-seed runs.
