+++ neat/read_simulator/utils/stitch_outputs.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import argparse
 import gzip
+import re
 import shutil
 import subprocess
 import sys
@@ -13,14 +14,12 @@ __all__ = ["main"]
 
 
 def print_stderr(msg: str, *, exit_: bool = False):
-    """Print to stderr, with an option to exit."""
     print(msg, file=sys.stderr)
     if exit_:
         sys.exit(1)
 
 
 def gather(paths: Iterable[Path], suffixes: tuple[str, ...]) -> List[Path]:
-    """Gather all files in the provided path whose name ends with the given suffix."""
     out: list[Path] = []
     for p in paths:
         if p.is_file() and p.name.endswith(suffixes):
@@ -48,7 +47,6 @@ def concat(files: List[Path], dest: Path, keep_compression: bool):
                 with f.open("rb") as in_f:
                     shutil.copyfileobj(in_f, out_f)
     else:
-        # stream out uncompressed lines for speed
         with dest.open("wb") as out_f:
             for f in files:
                 opener = gzip.open if is_gzipped(f) else open
@@ -67,7 +65,7 @@ def merge_bam(bams: List[Path], dest: Path, samtools: str):
     subprocess.check_call([samtools, "sort", "-o", str(dest), str(unsorted)])
     subprocess.check_call([samtools, "index", str(dest)])
     unsorted.unlink(missing_ok=True)
-
+    normalize_bam_header(dest, samtools)
 
 def merge_vcf(vcfs: List[Path], dest: Path):
     if not vcfs:
@@ -83,36 +81,134 @@ def merge_vcf(vcfs: List[Path], dest: Path):
                         out_f.write(line)
 
 
+def normalize_bam_header(bam: Path, samtools: str):
+    """Make BAM header deterministic: drop @PG, sort @SQ/@RG."""
+    hdr_text = subprocess.check_output([samtools, "view", "-H", str(bam)], text=True)
+
+    hd = []
+    sq = []
+    rg = []
+    other = []
+    for line in hdr_text.splitlines():
+        if line.startswith("@PG"):
+            continue
+        if line.startswith("@HD"):
+            hd.append(line)
+        elif line.startswith("@SQ"):
+            sq.append(line)
+        elif line.startswith("@RG"):
+            # optional: drop volatile CL: field
+            parts = [t for t in line.split("\t") if not t.startswith("CL:")]
+            line = "\t".join(parts)
+            rg.append(line)
+        else:
+            other.append(line)
+
+    def sn(x):
+        for t in x.split("\t"):
+            if t.startswith("SN:"):
+                return t[3:]
+        return x
+
+    def rgid(x):
+        for t in x.split("\t"):
+            if t.startswith("ID:"):
+                return t[3:]
+        return x
+
+    sq.sort(key=sn)
+    rg.sort(key=rgid)
+
+    new_hdr = "\n".join([*(hd or ["@HD\tVN:1.6\tSO:coordinate"]), *sq, *rg, *other]) + "\n"
+    hdr_tmp = bam.with_suffix(".reheader.sam")
+    out_tmp = bam.with_suffix(".reheader.bam")
+    hdr_tmp.write_text(new_hdr)
+
+    with out_tmp.open("wb") as out_f:
+        subprocess.check_call([samtools, "reheader", str(hdr_tmp), str(bam)], stdout=out_f)
+
+    out_tmp.replace(bam)
+    hdr_tmp.unlink(missing_ok=True)
+    subprocess.check_call([samtools, "index", str(bam)])
+
+
+def natural_key(path: Path):
+    s = path.name
+    parts = re.split(r"(\d+)", s)
+    return [int(p) if p.isdigit() else p for p in parts]
+
+
[…]
