+++ neat/read_simulator/utils/parallelize.py
@@ -1,3 +1,6 @@
+# TO DO: Sub-command in neat/cli/commands (make new file - move argparse there) - run as a separate command
+# TO DO: Try using a separate parallelization config
+
 from __future__ import annotations
 
 import argparse
@@ -33,21 +36,28 @@ def parse_args(argv: List[str] | None = None):
     )
 
     p.add_argument("config", type=Path, help="NEAT YAML/YML config containing the 'reference:' field")
-    p.add_argument("--outdir", type=Path, required=True, help="Top‑level directory to hold splits, per‑chunk outputs, and final stitched results",)
+    p.add_argument("--outdir", type=Path, required=True,
+                   help="Top-level directory to hold splits, per-chunk outputs, and final stitched results",)
 
     split = p.add_argument_group("splitting options")
     split.add_argument("--by", choices=["contig", "size"], default="contig", help="Split mode")
     split.add_argument("--size", type=int, default=1_000_000, help="Target chunk size when --by size")
-    split.add_argument("--cleanup-splits", action="store_true", help="Delete the 'splits' directory after stitching completes")
-    split.add_argument("--reuse-splits", action="store_true", help="Skip the splitting step and reuse any YAML/FASTA files already present in the 'splits' folder")
+    split.add_argument("--cleanup-splits", action="store_true",
+                       help="Delete the 'splits' directory after stitching completes")
+    split.add_argument("--reuse-splits", action="store_true",
+                       help="Skip the splitting step and reuse any YAML/FASTA files already present in the 'splits' folder")
 
     sim = p.add_argument_group("simulation options")
-    sim.add_argument("--jobs", type=int, default=os.cpu_count() or 2, help="Maximum number of parallel NEAT jobs to run",)
-    sim.add_argument("--neat-cmd", default="python -m neat.read_simulator.run", help="Command used to launch the read simulator (will be split on whitespace)",)
+    sim.add_argument("--jobs", type=int, default=os.cpu_count() or 2,
+                     help="Maximum number of parallel NEAT jobs to run",)
+    sim.add_argument("--neat-cmd", default="python -m neat.read_simulator.run",
+                     help="Command used to launch the read simulator (will be split on whitespace)",)
 
     stitch = p.add_argument_group("stitching options")
-    stitch.add_argument("--samtools", default="samtools", help="Path to samtools executable used by stitch_outputs.py",)
-    stitch.add_argument("--final-prefix", type=Path, default=Path("stitched/final"), help="Prefix (no extension) for the stitched outputs",)
+    stitch.add_argument("--samtools", default="samtools",
+                        help="Path to samtools executable used by stitch_outputs.py",)
+    stitch.add_argument("--final-prefix", type=Path, default=Path("stitched/final"),
+                        help="Prefix (no extension) for the stitched outputs",)
 
     return p.parse_args(argv)
 
@@ -76,14 +86,11 @@ def main(argv: List[str] | None = None):
         print(f"'reference:' not found or file does not exist: {ref_path}", file=sys.stderr)
         sys.exit(1)
 
-    # 1. Split the input reference
-
+    # 1) Split (or reuse)
     need_split = not args.reuse_splits
-
-    if args.reuse_splits and list((args.outdir / "splits").glob("*.y*ml")):
+    if args.reuse_splits and list(splits_dir.glob("*.y*ml")):
         print("[parallel] Recycling existing split FASTA/YAML files – skipping split step")
-    else:
[…]

[…]
