### 2025-08-23 3904c88
Command neat parallel implementation before current edits.
