+++ config_template/template_parallel_config.yml
@@ -1,0 +1,46 @@
+## Template for gen_reads parallel
+## Any parameter that is not required but has a default value will use the
+## default value even if the variable is not included in the config. For
+## required items, they must be included in the config and the must be given a value.
+## All other items can be present or not. If present and the value is set to a single
+## period, the variable will be treated as though it had been omitted. Please do
+## not modify this template, but instead make a copy in your working directory. Done this
+## way, you can run without even needing to declare additional flags.
+
+# Parallelization controls (used by neat parallel)
+
+# Top-level output directory for splits, per-chunk outputs, and stitched results.
+# type = string | required: yes
+outdir: REQUIRED
+
+# Location (prefix, no extension) for stitched outputs relative to outdir
+# type = string | required: no | default = stitched/final
+final_prefix: .
+
+# How to split the input reference for parallelization
+# type = string | required: no | default = contig | values: contig, size
+by: .
+
+# Target chunk size if by=size (overlap is derived from read_len*2)
+# type = int | required: no | default = 1000000
+size: .
+
+# Maximum number of concurrent NEAT jobs
+# type = int | required: no | default = (CPU count)
+jobs: .
+
+# Command used to launch the simulator
+# type = string | required: no | default = "neat read-simulator"
+neat_cmd: .
+
+# Path to samtools (name is fine if it's on PATH)
+# type = string | required: no | default = samtools
+samtools: .
+
+# If true, delete the 'splits' folder after stitching
+# type = bool | required: no | default = false
+cleanup_splits: .
+
+# If true, reuse existing files in 'splits' and skip the split step
+# type = bool | required: no | default = false
+reuse_splits: .
