+++ config_template/simple_parallel_template.yml
@@ -1,0 +1,9 @@
+outdir: REQUIRED
+final_prefix: .
+by: .
+size: .
+jobs: .
+neat_cmd: .
+samtools: .
+cleanup_splits: .
+reuse_splits: .
