### 2025-08-29 49eca2e
Partial progress on config errors.
