+++ neat/read_simulator/utils/parallelize.py
@@ -35,43 +35,31 @@ def parse_args(argv: List[str] | None = None):
     )
 
     p.add_argument("config", type=Path, help="NEAT YAML/YML config containing the 'reference:' field")
-    p.add_argument("--outdir", type=Path, required=True,
-                   help="Top-level directory to hold splits, per-chunk outputs, and final stitched results",)
+    p.add_argument("--outdir", type=Path, required=False, default=None,
+                   help="Top-level directory for splits and stitched results (optional)")
 
     split = p.add_argument_group("splitting options")
-    split.add_argument("--by", choices=["contig", "size"], default="contig", help="Split mode")
-    split.add_argument("--size", type=int, default=1_000_000, help="Target chunk size when --by size")
-    split.add_argument("--cleanup-splits", action="store_true",
+    split.add_argument("--by", choices=["contig", "size"], default=None, help="Split mode")
+    split.add_argument("--size", type=int, default=None, help="Target chunk size when --by size")
+    split.add_argument("--cleanup-splits", action=argparse.BooleanOptionalAction, default=None,
                        help="Delete the 'splits' directory after stitching completes")
-    split.add_argument("--reuse-splits", action="store_true",
-                       help="Skip the splitting step and reuse any YAML/FASTA files already present in the 'splits' folder")
+    split.add_argument("--reuse-splits", action=argparse.BooleanOptionalAction, default=None,
+                       help="Skip the splitting step and reuse any YAML/FASTA files already present")
 
     sim = p.add_argument_group("simulation options")
-    sim.add_argument("--jobs", type=int, default=os.cpu_count() or 2,
-                     help="Maximum number of parallel NEAT jobs to run",)
-    sim.add_argument("--neat-cmd", default="python -m neat.read_simulator.run",
-                     help="Command used to launch the read simulator (will be split on whitespace)",)
+    sim.add_argument("--jobs", type=int, default=None, help="Maximum number of parallel NEAT jobs to run")
+    sim.add_argument("--neat-cmd", default=None, help="Command to launch the read simulator")
 
     stitch = p.add_argument_group("stitching options")
-    stitch.add_argument("--samtools", default="samtools",
-                        help="Path to samtools executable used by stitch_outputs.py",)
-    stitch.add_argument("--final-prefix", type=Path, default=Path("stitched/final"),
-                        help="Prefix (no extension) for the stitched outputs",)
+    stitch.add_argument("--samtools", default=None, help="Path to samtools executable used by stitch_outputs.py")
+    stitch.add_argument("--final-prefix", type=Path, default=None, help="Prefix for stitched outputs")
 
     return p.parse_args(argv)
 
 
 def main(argv: List[str] | None = None):
+
     args = parse_args(argv)
-
-    args.outdir.mkdir(parents=True, exist_ok=True)
-    splits_dir = args.outdir / "splits"
-    sims_dir = args.outdir / "sim_runs"
-    splits_dir.mkdir(parents=True, exist_ok=True)
-    sims_dir.mkdir(parents=True, exist_ok=True)
-
-    final_prefix = args.final_prefix if args.final_prefix.is_absolute() else args.outdir / args.final_prefix
-
     try:
         import yaml
     except ModuleNotFoundError:
@@ -79,11 +67,119 @@ def main(argv: List[str] | None = None):
         sys.exit(1)
 
     cfg_dict = yaml.safe_load(args.config.read_text())
[…]
