### 2025-09-03 6c0d9c3
Parallelization completed paper.md
