+++ neat/parallel_read_simulator/split_inputs.py
@@ -1,5 +1,5 @@
 """
-Description.
+Split a FASTA and NEAT config into per‑contig or fixed‑size chunks ready for simulation.
 """
 
 import argparse
@@ -9,13 +9,22 @@ import yaml
 import time
 from pathlib import Path
 from textwrap import wrap
-from typing import Iterator, List
+from typing import Iterator, List, Tuple
+
+import logging
 
 __all__ = ["main"]
 
+_LOG = logging.getLogger(__name__)
+
 
 class SimpleRecord:
-    """Basic stand-in for Bio.SeqRecord.SeqRecord."""
+    """
+    Basic stand‑in for Bio.SeqRecord.SeqRecord.
+    This lightweight class stores an identifier and a sequence and
+    implements ``__len__`` so that sequence length can be computed.
+    """
+
     __slots__ = ("id", "seq")
 
     def __init__(self, id_: str, seq: str):
@@ -27,8 +36,9 @@ class SimpleRecord:
 
 
 # Utility helpers
-def print_stderr(msg: str, *, exit_: bool = False):
-    print(msg, file=sys.stderr)
+def print_stderr(msg: str, *, exit_: bool = False) -> None:
+    """Log an error message and optionally exit with status 1."""
+    _LOG.error(msg)
     if exit_:
         sys.exit(1)
 
@@ -39,6 +49,7 @@ def disk_bytes_free(path: Path) -> int:
 
 # FASTA helpers
 def parse_fasta(path: Path) -> Iterator[SimpleRecord]:
+    """Parse a FASTA file into SimpleRecord objects."""
     with path.open() as fh:
         header: str | None = None
         seq_chunks: list[str] = []
@@ -58,6 +69,7 @@ def parse_fasta(path: Path) -> Iterator[SimpleRecord]:
 
 
 def write_fasta(rec: SimpleRecord, out_fa: Path, width: int = 60) -> None:
+    """Write a sequence record to a FASTA file, wrapping lines at a given width."""
     out_fa.parent.mkdir(parents=True, exist_ok=True)
     with out_fa.open("w") as fh:
         fh.write(f">{rec.id}\n")
@@ -66,8 +78,10 @@ def write_fasta(rec: SimpleRecord, out_fa: Path, width: int = 60) -> None:
 
 
 # Chunk generators
-def chunk_record(record: SimpleRecord, chunk_len: int, overlap: int):
-    """Yield overlapping chunk_len slices."""
+def chunk_record(record: SimpleRecord, chunk_len: int, overlap: int) -> Iterator[Tuple[SimpleRecord, int]]:
+    """
+    Yield overlapping slices of a record.
+    """
     seq_len = len(record)
     start = 0
     chunk_id = 1
@@ -82,7 +96,8 @@ def chunk_record(record: SimpleRecord, chunk_len: int, overlap: int):
         chunk_id += 1
 
 
-def write_config(template_cfg: dict, out_fa: Path, out_yml: Path):
+def write_config(template_cfg: dict, out_fa: Path, out_yml: Path) -> None:
+    """Write a per‑fragment configuration derived from the template config."""
     cfg_copy = template_cfg.copy()
     cfg_copy["reference"] = str(out_fa.resolve())
     if "name" in cfg_copy:
@@ -91,7 +106,7 @@ def write_config(template_cfg: dict, out_fa: Path, out_yml: Path):
 
 
 # Argument parsing
-def parse_args(argv: List[str] | None = None):
+def parse_args(argv: List[str] | None = None) -> argparse.Namespace:
     p = argparse.ArgumentParser(
         description="Split a FASTA and NEAT config into per-contig or fixed-size chunks ready for neat read-simulator.",
[…]
