### 2025-09-22 #8375 @bricoletc
@csoneson happy to review this as well!
