### 2025-09-26 #9056 @erik-whiting
My review notes for the paper part only:

Overall: Paper talks about version 4.2 but the repo is on 4.3 it looks like. Would it be easy to update the paper to talk about 4.3 or do we only want to introduce 4.2 with this paper?

Summary: The summary doesn’t introduce the software at all, just gives an overview of some computational challenges around working with NGS data. Suggestion: Please at least briefly introduce NEAT here/

Statement of Need: The statement of need section is more like a description of the changes made in this version of NEAT. You don’t talk about why this software suite is needed or how it compares to similar tools in this space.

State of the Field: Missing

Quality of writing is great and the references seem to be in order
