### 2025-10-01 ncsa/NEAT#161 @chenyenchung
ValueError when VCF quality score is missing ('.')
**Describe the bug**
When using an input VCF file with missing quality scores (represented as `'.'` in the QUAL column), NEAT crashes with `ValueError: invalid literal for int() with base 10: '.'` during read generation.

**To Reproduce**

Use the following config:

```
reference: ./H1N1.fasta
read_len: 50
paired_ended: True
fragment_mean: 100
fragment_st_dev: 10
include_vcf: ./H1N1_min.vcf
```

With the variant file (`H1N1_min.vcf`) containing:

```
##fileformat=VCFv4.2
##reference=file://H1N1.fa
#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	NA00001
H1N1_HA	6	.	G	A	.	.	AF=10;PP=11	.	.
```

`neat read-simulator -c neat_config.yml -o simulated_data` will fail during mutation generation:

```
2025-10-01 20:41:21,947:INFO:neat.read_simulator.utils.generate_variants:Added 2 mutations to H1N1_HA
2025-10-01 20:41:21,947:INFO:neat.read_simulator.utils.generate_reads:Sampling reads...
2025-10-01 20:41:21,959:ERROR:neat:read-simulator failed, see the traceback below
Traceback (most recent call last):
  File "/home/mambauser/NEAT/neat/cli/cli.py", line 131, in main
    cmd(args)
  File "/home/mambauser/NEAT/neat/cli/commands/read_simulator.py", line 47, in execute
    read_simulator_runner(arguments.config, arguments.output)
  File "/home/mambauser/NEAT/neat/read_simulator/runner.py", line 316, in read_simulator_runner
    read1_fastq_paired, read1_fastq_single, read2_fastq_paired, read2_fastq_single = generate_reads(
  File "/home/mambauser/NEAT/neat/read_simulator/utils/generate_reads.py", line 345, in generate_reads
    read_1.finalize_read_and_write(
  File "/home/mambauser/NEAT/neat/read_simulator/utils/read.py", line 343, in finalize_read_and_write
    self.read_quality_string = "".join([chr(int(x) + quality_offset) for x in self.quality_array])
  File "/home/mambauser/NEAT/neat/read_simulator/utils/read.py", line 343, in <listcomp>
    self.read_quality_string = "".join([chr(int(x) + quality_offset) for x in self.quality_array])
ValueError: invalid literal for int() with base 10: '.'
ERROR: read-simulator failed, showing the last error
Traceback (most recent call last):
  File "/home/mambauser/NEAT/neat/cli/cli.py", line 131, in main
    cmd(args)
  File "/home/mambauser/NEAT/neat/cli/commands/read_simulator.py", line 47, in execute
    read_simulator_runner(arguments.config, arguments.output)
  File "/home/mambauser/NEAT/neat/read_simulator/runner.py", line 316, in read_simulator_runner
    read1_fastq_paired, read1_fastq_single, read2_fastq_paired, read2_fastq_single = generate_reads(
  File "/home/mambauser/NEAT/neat/read_simulator/utils/generate_reads.py", line 345, in generate_reads
    read_1.finalize_read_and_write(
  File "/home/mambauser/NEAT/neat/read_simulator/utils/read.py", line 343, in finalize_read_and_write
    self.read_quality_string = "".join([chr(int(x) + quality_offset) for x in self.quality_array])
  File "/home/mambauser/NEAT/neat/read_simulator/utils/read.py", line 343, in <listcomp>
    self.read_quality_string = "".join([chr(int(x) + quality_offset) for x in self.quality_array])
ValueError: invalid literal for int() with base 10: '.'
```

**Expected behavior**
Missing QUAL ('.') in VCF files to be replaced with a default value.
