### 2025-10-02 07b99fa
Handling missing QUAL in user VCF files
When QUAL is missing in user VCF files, replace it with 42.
(Address #161)
