### 2025-10-02 #9056 @chenyenchung
Hi @keshav-gandhi, congratulations on the great work. Please find below my review with a few suggestions and questions:

## Overall

I like the brevity and clarity of how the manuscript was written, and find the installation and usage instruction in the repository easy to follow. However, I found it not immediately obvious what the tool was designed for due to the rapid expansion of next-generation sequencing applications.

- Suggestion 1: Given the main use case appears to be simulation of short-read whole-genome or whole-exome sequencing, can this be explicit in the title, summary, and statement of need?

## Functionality

The instruction is fine and streamlined as it is, but I have similar confusion as @bricoletc:

- Question 1: Why are both `poetry` and `conda` necessary (are they)?
- Question 2: Is the `bioconda` package working properly? If so, it could be a more user-friendly way of distribution.

And, given the current `environment.yml`, due to the specification of `libgcc` version, the installation can only be done per instruction on Linux but not MacOS (I don't have access to a Windows machine, so I don't know if it works on Windows and WSL).

- Suggestion 2: Please specify the platform you intend to support.

## Documentation

I like a lot about the example configuration file (`config_template/template_neat_config.yml`) having instruction baked in, and this is almost the only thing that I need after installation to get things up and running.

- Suggestion 3: While not really hard to find, mentioning the existence of the template in `README.md` would be nice.

Automatic tests are great and incorporated in GitHub Action, but some features could benefit from more coverage. For example, using the example data to run a minimal example caused a crash (https://github.com/ncsa/NEAT/issues/161). Two pull requests (https://github.com/ncsa/NEAT/pull/162 and https://github.com/ncsa/NEAT/pull/163) were submitted and merged to address this issue.

- Suggestion 4: A higher-level integration test that take the example data for a quick run could be nice.

In addition, https://github.com/ncsa/NEAT/pull/162 got me into noticing that as a simulation tool, how error models and mutation models work appear not to have been documented beyond [the original release manuscript](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0167047), but given Table 1, I assume these have changed in the current version submitted for review. Therefore,

- Suggestion 5: How error models and mutation models work should be documented either in the manuscript or `README.md`.
- Suggestion 6: The error model and mutation models should be tested in the test suite. This would also help clarifying whether some behavior is intended. For example, the semantics of `how_many` appears to indicate that we are expecting a given number of alternative alleles to be present, but in [neat/common/ploid_functions.py](https://github.com/ncsa/NEAT/blob/60ef37aacb5c671b31fcd31a9e52f75cafb8a590/neat/common/ploid_functions.py#L49), the implementation seems to generate _at most_ `how_many` alternative alleles since line 47 can select the same chromosome multiple times. I understand that whether a chromosome can be sampled with replacement is not right nor wrong, but it would be nice to have a clear intention so the users know what to expect if they wish to understand what was simulated with `NEAT`.

## Others

- Suggestion 7: A discussion of tools for the same or similar purposes is currently lacking. It could be nice to include the tools discussed in Alosaimi _et al._ and a functionality matrix to show case the feature set of these tools.
