### 2025-10-04 8aa2032
Developed script for wrapping bacterial chromosomes
