+++ BacterialWrapperScript/Outputs/Salmonella/salmonella_config_test.yml
@@ -1,0 +1,137 @@
+# Test 1: Default parameters, Salmonella data
+
+## Template for gen_reads parallel
+## Any parameter that is not required but has a default value will use the
+## default value even if the variable is not included in the config. For
+## required items, they must be included in the config and the must be given a value.
+## All other items can be present or not. If present and the value is set to a single
+## period, the variable will be treated as though it had been omitted. Please do
+## not modify this template, but instead make a copy in your working directory. Done this
+## way, you can run without even needing to declare -c.
+
+# Absolute path to input reference fasta file
+# type = string | required: yes
+reference: BacterialWrapperScriptReferenceFiles/Salmonella/GCF_000006945.2_ASM694v2_genomic.fna
+
+# Read length of the reads in the fastq output. Only required if @produce_fastq is set to true
+# type = int | required: no | default = 101
+read_len: .
+
+# Number of threads to request for NEAT. The recommended amount is the number of chromosomes in
+# your input fasta plus 1.
+# type = int | required: no | default = 1
+threads: .
+
+# Average Coverage for the entire genome.
+# type = float | required: no | default = 10.0
+coverage: .
+
+# Absolute path to file with sequencing error model
+# type = string | required: no | default: <NEAT_DIR>/neat/models/defaults/default_error_model.pickle.gz
+error_model: .
+
+# Average sequencing error rate for the sequencing machine
+# type = float | required = no | must be between 0.0 and 0.3
+avg_seq_error: .
+
+# This scales the quality scores to match the desired average sequencing error rate
+# specified by avg_seq_error.
+# type: boolean | required = no | default = false
+rescale_qualities: .
+
+# This is the factor to add to the quality scores to get the ascii text version of the
+# score. The default follows the sanger quality offset
+# type: int | required = no | default = 33
+quality_offset: .
+
+# Desired ploidy
+# type = int | required = no | default = 2
+ploidy: .
+
+# Absolute path to vcf file containing variants that will always be included, regardless
+# of genotype and filter. You can pre-filter your vcf for these fields before inputting it
+# if this is not the desired behavior.
+# type: string | required = no
+input_variants: .
+
+# Absolute path to bed file containing reference regions that the simulation
+# should target.
+# type = string | required = no
+target_bed: .
+
+# Absolute path to bed file containing reference regions that the simulation
+# should discard.
+# type = string | required = no
+discard_bed: .
+
+# Absolute path to the mutation model pickle file. Omitting this value will cause
+# NEAT to use the default model, with some standard parameters, and generally uniform biases.
+# type: string | required = no
+mutation_model: .
+
+# Average mutation rate per base pair. Overall average is 0.001, or model default
+# Use either this value to override the mutation rate for the default or input model.
+# type: float | required = no | must be between 0.0 and 0.3
+mutation_rate: .
+
+# Absolute path to a bed file with mutation rates by region.
+# Rates must be in the fourth column and be of the form "mut_rate=x.xx"
+# Rates must be between 0.00 and 0.03
+# type: string | required = no
+mutation_bed: .
+
+# Whether the output should be paired ended. For certain conditions (i.e., vcf only or
+# fasta only), this will be ignored. If this is true, then there must be an included fragment
+# length model output from runner.py or a mean and standard deviation
+# by declaring values for @fragment_mean and @fragment_std_dev.
+# type: boolean | required = no | default = false
+paired_ended: .
+
+# Absolute path to a pickle file containing the fragment length model output
+# from runner.py.
+# type: string | required = no | default: <NEAT_DIR>/neat/models/defaults/default_fraglen_model.pickle.gz
+fragment_model: .
+
+# Mean for the paired end fragment length. This only applies if paired-ended is set to true.
[…]
