### 2025-10-04 383a4ef
Developed script for wrapping bacterial chromosomes
