### 2025-10-04 7ca87c9
Updating parallel script to call functions in neat directly instead of using subscripts
