+++ neat/read_simulator/runner.py
@@ -1,124 +1,27 @@
 """
 Runner for generate_reads task
 """
-import time
 import logging
-import pickle
-import gzip
 
-from Bio import SeqIO
 from pathlib import Path
 
-from .utils import Options, parse_input_vcf, parse_beds, OutputFileWriter, \
-    generate_variants, generate_reads
-from ..common import validate_input_path, validate_output_path
-from ..models import MutationModel, SequencingErrorModel, FragmentLengthModel, TraditionalQualityModel
-from ..models.default_cancer_mutation_model import *
-from ..variants import ContigVariants
+from .utils import Options
+from ..common import validate_input_path
+from .parallel_runner import main as parallel_runner
+from .single_runner import read_simulator_single as single_runner
 
 __all__ = ["read_simulator_runner"]
 
 _LOG = logging.getLogger(__name__)
 
-
-def initialize_all_models(options: Options):
-    """
-    Helper function that initializes models for use in the rest of the program.
-    This includes loading the model and attaching the rng for this run
-    to each model, so we can perform the various methods.
-
-    :param options: the options for this run
-    """
-
-    # Load mutation model or instantiate default
-    if options.mutation_model:
-        mut_model = pickle.load(gzip.open(options.mutation_model))
-    else:
-        mut_model = MutationModel()
-
-    # Set random number generator for the mutations:
-    mut_model.rng = options.rng
-    # Set custom mutation rate for the run, or set the option to the input rate so we can use it later
-    if options.mutation_rate is not None:
-        mut_model.avg_mut_rate = options.mutation_rate
-
-    cancer_model = None
-    if options.cancer and options.cancer_model:
-        # cancer_model = pickle.load(gzip.open(options.cancer_model))
-        # Set the rng for the cancer mutation model
-        cancer_model.rng = options.rng
-    elif options.cancer:
-        # Note all parameters not entered here use the mutation madel defaults
-        cancer_model = MutationModel(
-            avg_mut_rate=default_cancer_avg_mut_rate,
-            homozygous_freq=default_cancer_homozygous_freq,
-            variant_probs=default_cancer_variant_probs,
-            insert_len_model=default_cancer_insert_len_model,
-            is_cancer=True
-        )
-
-    _LOG.debug("Mutation models loaded")
-
-    # We need sequencing errors to get the quality score attributes, even for the vcf
-    if options.error_model:
-        error_models = pickle.load(gzip.open(options.error_model))
-        error_model_1 = error_models["error_model1"]
-        quality_score_model_1 = error_models["qual_score_model1"]
-        if options.paired_ended:
-            if error_models["error_model2"]:
-                error_model_2 = error_models["error_model2"]
-                quality_score_model_2 = error_models["qual_score_model2"]
-            else:
-                _LOG.warning('Paired ended mode declared, but input sequencing error model is single ended,'
-                             'duplicating model for both ends')
-                error_model_2 = error_models["error_model1"]
-                quality_score_model_2 = error_models["qual_score_model1"]
-        else:
-            # ignore second model if we're in single-ended mode
-            error_model_2 = None
-            quality_score_model_2 = None
-    else:
-        # Use all the default values
-        error_model_1 = SequencingErrorModel()
-        quality_score_model_1 = TraditionalQualityModel()
-        if options.paired_ended:
-            error_model_2 = SequencingErrorModel()
-            quality_score_model_2 = TraditionalQualityModel()
-        else:
-            error_model_2 = None
-            quality_score_model_2 = None
-
-    _LOG.debug('Sequencing error and quality score models loaded')
-
-    if options.fragment_model:
[…]
