+++ neat/cli/commands/options.py
@@ -9,9 +9,18 @@ from .base import Group
 output_group = Group("output", is_mutually_exclusive=True, required=True)
 output_group.add_argument(
     "-o",
-    "--output",
-    dest="output",
+    "--output_dir",
+    dest="output_dir",
     type=str,
-    help="Path (including filename prefix) to the output file",
+    help="Path to the output directory. Will create if not present.",
     default=None
+)
+
+output_group.add_argument(
+    "-p",
+    "--prefix",
+    dest="prefix",
+    type=str,
+    help="Prefix to use to name files",
+    default="neat_sim"
 )
