### 2025-10-04 8ec8e23
Reworking parallelism to integrate it into NEAT more seamlessly
