### 2025-10-05 5bda226
fixed bam creation
