+++ neat/read_simulator/utils/generate_variants.py
@@ -13,6 +13,7 @@ import sys
 from Bio import SeqRecord
 from numpy.random import Generator
 
+from . import OptionsPerThread
 from ...models import MutationModel
 from ...variants import Insertion, Deletion, SingleNucleotideVariant, ContigVariants
 from ..utils import Options
@@ -57,7 +58,7 @@ def generate_variants(
         mutation_rate_regions: list[tuple[int, int]],
         existing_variants: ContigVariants,
         mutation_model: MutationModel,
-        options: Options,
+        options: OptionsPerThread,
         max_qual_score: int
 ) -> ContigVariants:
     """
