+++ neat/cli/commands/options.py
@@ -4,6 +4,8 @@ Definitions of shared subcommand options.
 
 __all__ = ["output_group"]
 
+import os
+
 from .base import Group
 
 output_group = Group("output", is_mutually_exclusive=True, required=True)
@@ -13,7 +15,7 @@ output_group.add_argument(
     dest="output_dir",
     type=str,
     help="Path to the output directory. Will create if not present.",
-    default=None
+    default=os.getcwd()
 )
 
 output_group.add_argument(
