+++ neat/read_simulator/utils/options.py
@@ -35,7 +35,6 @@ from ...common import validate_input_path, validate_output_path
 
 _LOG = logging.getLogger(__name__)
 
-
 class Options(SimpleNamespace):
     """
     class representing the options
@@ -47,7 +46,8 @@ class Options(SimpleNamespace):
     :param rng_seed: A seed to use for a test run (for reproducible tests)
     """
     def __init__(self,
-                 output_path: Path = None,
+                 output_path: Path,
+                 output_prefix: str,
                  config_file: Path = None,
                  reference: Path = None,
                  rng_seed: int = None
@@ -78,7 +78,9 @@ class Options(SimpleNamespace):
         # TODO maybe redo as a dataclass?
         self.defs = {}
         self.config_file = config_file
-        self.output = output_path
+        self.output_dir = output_path
+        self.output_prefix = output_prefix
+        self.output = self.output_dir / self.output_prefix
 
         self.defs['reference'] = (str, reference, 'exists', None)
         self.defs['read_len'] = (int, 151, 10, inf)
@@ -94,11 +96,6 @@ class Options(SimpleNamespace):
         self.defs['mutation_rate'] = (float, None, 0, 0.3)
         self.defs['mutation_bed'] = (str, None, 'exists', None)
         self.defs['quality_offset'] = (int, 33, 33, 64)
-
-        # Params for cancer (not implemented yet)
-        self.defs['cancer'] = (bool, False, None, None)
-        self.defs['cancer_model'] = (str, None, 'exists', None)
-        self.defs['cancer_purity'] = (float, 0.8, 0.0, 1.0)
 
         self.defs['paired_ended'] = (bool, False, None, None)
         self.defs['fragment_model'] = (str, None, 'exists', None)
@@ -118,12 +115,12 @@ class Options(SimpleNamespace):
         self.defs['parallel_mode'] = (bool, False, None, None)
         self.defs['mode'] = (str, 'size', 'choice', ['size', 'contig'])
         self.defs['size'] = (int, 500000, None, None)
-        self.defs['threads'] = (int, 999, 1, 1000)
+        self.defs['threads'] = (int, 1, 1, 1000)
         self.defs['cleanup_splits'] = (bool, False, None, None)
         self.defs['reuse_splits'] = (str, None, 'exists', None)
 
         # Create base variables, for update by the config
-        self.reference: str | Path = reference
+        self.reference: Path | None = reference
         self.read_len: int = 151
         self.threads: int = 1
         self.coverage: int = 10
@@ -158,11 +155,6 @@ class Options(SimpleNamespace):
         self.overwrite_output: bool = False
         self.rng_seed: int | None = None
         self.rng: Generator | None = None
-
-        # Cancer options (not yet implemented)
-        self.cancer: str | Path
-        self.cancer_model: bool
-        self.cancer_purity: float
 
         # These parameters relate to parallelization
         self.parallel_mode: bool = False
@@ -215,7 +207,7 @@ class Options(SimpleNamespace):
             if value_to_check not in crit2:
                 _LOG.error(f"Must choose one of {crit2}")
                 sys.exit(1)
-        elif crit1.isdigit() and crit2.isdigit():
+        elif isinstance(crit1, int) and isinstance(crit2, int):
             if not (crit1 <= value_to_check <= crit2):
                 _LOG.error(f'`{keyname}` must be between {crit1} and {crit2} (input: {value_to_check}).')
                 sys.exit(1)
@@ -225,7 +217,6 @@ class Options(SimpleNamespace):
         This sets up the option attributes. It's not perfect, because it sort of kills type hints.
         But I'm not sure how else to accomplish this.
         """
-        # Skip trying to read the config for a test run
[…]
