+++ neat/read_simulator/parallel_runner.py
@@ -3,18 +3,14 @@ Split a reference genome, run NEAT simulations in
 parallel, and stitch their outputs back together.
 """
 
-import argparse
-import os
-import shutil
 import subprocess
 import sys
-import tempfile
 import time
-from concurrent.futures import ProcessPoolExecutor, as_completed
 from pathlib import Path
 from typing import List, Tuple
+from multiprocessing import Pool
 
-from .utils import Options
+from .utils import Options, OptionsPerThread
 from .utils.split_inputs import main as split_main
 from .utils.stitch_outputs import main as stitch_main
 from .single_runner import read_simulator_single
@@ -22,6 +18,8 @@ from .single_runner import read_simulator_single
 import logging
 
 __all__ = ["main"]
+
+EXTENSIONS = ["gz", "fastq", "bam", "vcf"]
 
 _LOG = logging.getLogger(__name__)
 
@@ -58,78 +56,61 @@ def main(options: Options) -> None:
     splits_files = []
     if not need_split and list(options.existing_splits_dir.glob("*.y*ml")):
         _LOG.info("[parallel] Recycling existing split FASTA/YAML files – skipping split step")
+        split_sec = 0.0
     else:
         t0 = time.time()
         _LOG.info("[parallel] Splitting reference...")
         splits_files = split_main(options)
         split_sec = time.time() - t0
 
+    _LOG.setLevel(logging.ERROR)
     # 2) Submit NEAT sims in a deterministic order using the manifest
     _LOG.info(f"[parallel] Launching {len(splits_files)} NEAT job(s) (max {options.threads} in parallel)...")
     t1 = time.time()
-    futures = []
     output_opts = []
-    with ProcessPoolExecutor(max_workers=options.threads) as pool:
-        for splits_file in splits_files:
-            current_options = options.deep_copy()
-            current_options.reference = splits_file
-            current_output_dir = options.temp_dir / splits_file.stem
-            current_output_dir.mkdir(parents=True, exist_ok=True)
-            # Create local filenames based on fasta indexing scheme.
-            if options.produce_fastq:
-                current_options.fq1 = make_new_name(options.fq1, splits_file, current_output_dir)
-                if options.paired_ended:
-                    # if paired ended, there must be an fq2
-                    current_options.fq2 = make_new_name(options.fq2, splits_file, current_output_dir)
-            if options.produce_bam:
-                current_options.bam = make_new_name(options.bam, splits_file, current_output_dir)
-            if options.produce_vcf:
-                current_options.vcf = make_new_name(options.vcf, splits_file, current_output_dir)
-            output_opts.append(current_options)
-            current_options.output_dir = current_output_dir
-            futures.append(pool.submit(worker, read_simulator_single(current_options)))
+    for splits_file in splits_files:
+        reference = splits_file
+        current_output_dir = options.temp_dir_path / splits_file.stem
+        current_output_dir.mkdir(parents=True, exist_ok=True)
+        # Create local filenames based on fasta indexing scheme.
+        fq1 = None
+        fq2 = None
+        bam = None
+        vcf = None
+        if options.produce_fastq:
+            fq1 = current_output_dir / options.fq1.name
+            if options.paired_ended:
+                # if paired ended, there must be a fq2
+                fq2 = current_output_dir / options.fq2.name
+        if options.produce_bam:
+            bam = current_output_dir / options.bam.name
+        if options.produce_vcf:
+            vcf = current_output_dir / options.vcf.name
 
-        for fut in as_completed(futures):
-            try:
[…]
