### 2025-10-05 452a2eb
Removed unnecessary requirement in cli
