+++ neat/read_simulator/utils/options.py
@@ -13,6 +13,7 @@ We want to introduce a quick-run option as well that will allow command line onl
 file. To do so, we have a few possible inputs the init function can accept. Inputting any of these will
 trigger quick-run mode, setting most of the parameters to defaults.
 """
+from typing import Any
 
 import numpy as np
 import logging
@@ -45,20 +46,99 @@ class Options(SimpleNamespace):
     :param reference: Path to a reference for a test run
     :param rng_seed: A seed to use for a test run (for reproducible tests)
     """
+
     def __init__(self,
-                 output_path: Path,
+                 reference: Path | None = None,
+                 output_dir: Path | None = None,
+                 fq1: Path | None = None,
+                 fq2: Path | None = None,
+                 vcf: Path | None = None,
+                 bam: Path | None = None,
+                 output_prefix: str = "neat_sim",
+                 overwrite_output: bool = False,
+                 rng_seed: int | None = None,
+                 read_len: int = 101,
+                 threads: int = 1,
+                 coverage: int = 10,
+                 error_model: Path | None = None,
+                 avg_seq_error: float | None = None,
+                 rescale_qualities: bool = False,
+                 ploidy: int = 2,
+                 include_vcf: Path | None = None,
+                 target_bed: Path | None = None,
+                 discard_bed: Path | None = None,
+                 mutation_model: Path | None = None,
+                 mutation_rate: float | None = None,
+                 mutation_bed: Path | None = None,
+                 quality_offset: int = 33,
+                 paired_ended: bool = False,
+                 fragment_model: Path | None = None,
+                 fragment_mean: float | None = None,
+                 fragment_st_dev: float | None = None,
+                 produce_bam: bool = False,
+                 produce_vcf: bool = False,
+                 produce_fastq: bool = True,
+                 min_mutations: int = 0,
+                 parallel_mode: str = "contig",
+                 parallel_block_size: int = 500000,
+                 cleanup_splits: bool = True,
+                 splits_dir: Path | None = None,
+                 reuse_splits: bool = False,
+                 **kwargs: Any):
+        super().__init__(**kwargs)
+        self.reference: Path = reference
+        self.output_dir: Path = output_dir
+        self.output = self.output_dir / output_prefix
+        self.overwrite_output = overwrite_output
+        self.rng_seed = rng_seed
+        self.read_len: int = read_len
+        self.threads: int = threads
+        self.coverage: int = coverage
+        self.error_model: Path | None = error_model
+        self.avg_seq_error: float | None = avg_seq_error
+        self.rescale_qualities: bool = rescale_qualities
+        self.ploidy: int = ploidy
+        self.include_vcf: Path | None = include_vcf
+        self.target_bed: Path | None = target_bed
+        self.discard_bed: Path | None = discard_bed
+        self.mutation_model: Path | None = mutation_model
+        self.mutation_rate: float | None = mutation_rate
+        self.mutation_bed: str | None = mutation_bed
+        self.quality_offset: int = quality_offset
+
+        self.paired_ended: bool = paired_ended
+        self.fragment_model: str | None = fragment_model
+        self.fragment_mean: float | None = fragment_mean
+        self.fragment_st_dev: float | None = fragment_st_dev
+        self.produce_bam: bool = produce_bam
+        self.produce_vcf: bool = produce_vcf
+        self.produce_fastq: bool = produce_fastq
+        self.min_mutations: int = min_mutations
+
+        # Parallel features
+        self.parallel_mode: str = parallel_mode
+        self.parallel_block_size: int = parallel_block_size
+        self.cleanup_splits: bool = cleanup_splits
[…]
