+++ tests/test_models/test_seq_error.py
@@ -10,38 +10,33 @@ from Bio.SeqRecord import SeqRecord
 from Bio.Seq import Seq
 from importlib import resources
 
-from neat.models import SequencingErrorModel, default_sequencing_error_model
-from neat.variants import SingleNucleotideVariant, Insertion, Deletion
+from neat.models import SequencingErrorModel
+from neat.variants import SingleNucleotideVariant
 
 
-def test_get_seq_error_Insertion():
-    """Test that error_type code works correctly"""
+def test_get_seq_error_snv_only():
+    """With SNV-only probabilities, produced errors should be SNVs."""
     rng = np.random.default_rng(0)
-    error_model = SequencingErrorModel(read_length=10, rng=rng, variant_probs={Insertion: 1})
+    model = SequencingErrorModel(read_length=10, variant_probs={SingleNucleotideVariant: 1.0})
     quality_scores = np.full_like(np.arange(10), 36)
+    # Make at least one base very error-prone
     quality_scores[0] = 0
     reference = SeqRecord(Seq('AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA'), id="fake_1", name="fake", description="fake")
-    errors = error_model.get_sequencing_errors(reference, quality_scores)
-    assert errors[0][0] == Insertion
+    errors, _padding = model.get_sequencing_errors(padding=5, reference_segment=reference, quality_scores=quality_scores, rng=rng)
+    assert all(e.error_type == SingleNucleotideVariant for e in errors)
 
 
-def test_get_seq_error_Deletion():
-    """Test that error_type code works correctly"""
+essentially_zero = 0.0
+
+def test_no_errors_when_avg_zero():
+    """When average error is zero, the model should introduce no errors."""
     rng = np.random.default_rng(0)
-    error_model = SequencingErrorModel(read_length=10, rng=rng, variant_probs={Deletion: 1})
-    quality_scores = np.full_like(np.arange(10), 36)
-    quality_scores[0] = 0
-    reference = SeqRecord(Seq('AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA'), id="fake_1", name="fake", description="fake")
-    errors = error_model.get_sequencing_errors(reference, quality_scores)
-    assert errors[0][0] == Deletion
-
-
-def test_get_seq_error_SNV():
-    """Test that error_type code works correctly"""
-    rng = np.random.default_rng(0)
-    error_model = SequencingErrorModel(read_length=10, rng=rng, variant_probs={SingleNucleotideVariant: 1})
-    quality_scores = np.full_like(np.arange(10), 36)
-    quality_scores[0] = 0
-    reference = SeqRecord(Seq('AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA'), id="fake_1", name="fake", description="fake")
-    errors = error_model.get_sequencing_errors(reference, quality_scores)
-    assert errors[0][0] == SingleNucleotideVariant
+    model = SequencingErrorModel(read_length=10, avg_seq_error=essentially_zero)
+    qs = np.full_like(np.arange(10), 36)
+    ref = SeqRecord(Seq('AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA'), id="fake_1", name="fake", description="fake")
+    result = model.get_sequencing_errors(padding=5, reference_segment=ref, quality_scores=qs, rng=rng)
+    if isinstance(result, tuple):
+        errors, pad = result
+        assert errors == [] and pad == 5
+    else:
+        assert result == []
