+++ neat/read_simulator/utils/output_file_writer.py
@@ -21,7 +21,7 @@ from Bio import SeqIO
 from pathlib import Path
 from numpy.random import Generator
 
-from .options import OptionsPerThread
+from .options import Options
 from ...common import validate_output_path, open_output
 from .read import Read
 
@@ -72,7 +72,7 @@ class OutputFileWriter:
     :param bam_header: A dictionary of lengths of each contig from the reference, keyed by contig id.
     """
     def __init__(self,
-                 options: OptionsPerThread,
+                 options: Options,
                  bam_header: dict = None):
 
         # set the booleans

[…]
