### 2025-10-06 1d263fe
showing signs of life
