+++ neat/read_simulator/utils/output_file_writer.py
@@ -96,21 +96,21 @@ class OutputFileWriter:
         # Set up filenames based on booleans
         files_to_write = []
         if self.paired and self.write_fastq:
-            self.fastq1_fn = options.output.parent / f'{options.output.stem}_r1.fastq.gz'
-            self.fastq2_fn = options.output.parent / f'{options.output.stem}_r2.fastq.gz'
+            self.fastq1_fn = options.fq1
+            self.fastq2_fn = options.fq2
             self.fastq_fns = [self.fastq1_fn, self.fastq2_fn]
             files_to_write.extend(self.fastq_fns)
         elif self.write_fastq:
-            self.fastq1_fn = options.output.parent / f'{options.output.stem}.fastq.gz'
+            self.fastq1_fn = options.fq1
             self.fastq2_fn = self.temporary_dir / "dummy.fastq.gz"
             self.fastq_fns = [self.fastq1_fn, self.fastq2_fn]
             files_to_write.extend(self.fastq_fns)
         if self.write_bam:
-            self.bam_fn = options.output.parent / f'{options.output.stem}_golden.bam'
+            self.bam_fn = options.bam
             self.bam_keys = list(bam_header)
             files_to_write.append(self.bam_fn)
         if self.write_vcf:
-            self.vcf_fn = options.output.parent / f'{options.output.stem}_golden.vcf.gz'
+            self.vcf_fn = options.vcf
             files_to_write.append(self.vcf_fn)
 
         self.files_to_write = files_to_write
