### 2025-10-06 b24a6fe
Had to update a couple tests with some minor code changes
