+++ .github/workflows/python-app.yml
@@ -19,4 +19,5 @@ jobs:
 
       - name: Run unit tests (pytest)
         run: |
+          poetry install
           micromamba run -n test_neat pytest -q tests
