### 2025-10-06 15b4a86
Tests pass locally, just trying to fix the github yml file
