+++ .github/workflows/python-app.yml
@@ -19,5 +19,5 @@ jobs:
 
       - name: Run unit tests (pytest)
         run: |
-          poetry install
-          micromamba run -n test_neat pytest -q tests
+          micromamba run -n test_neat poetry install
+          micromamba run -n test_neat python -m neat pytest -q tests
