### 2025-10-06 7e938d1
pytest itself is proving elusive.
