### 2025-10-07 b97ea10
fixed problem with temp bam file names
