### 2025-10-10 982e3fd
Simplified the file writing portions. The way the code was working before, it required creating sub files which were then stitched back together into the output, but since parallel code ran on top of that, we were stitching 2 layers of files, but only running on a chunk at a time, so essentially there was a loop only running once. This cuts out some of the extra loops we were running. Still need to stitch final files and produce final VCF
