+++ neat/read_simulator/utils/output_file_writer.py
@@ -8,21 +8,15 @@ __all__ = [
     "OutputFileWriter"
 ]
 
-import gzip
 import re
-import shutil
-import time
 from struct import pack
 import logging
-import pickle
 
 from Bio import bgzf
-from Bio import SeqIO
 from pathlib import Path
-from numpy.random import Generator
 
 from .options import Options
-from ...common import validate_output_path, open_output
+from ...common import validate_output_path
 from .read import Read
 
 _LOG = logging.getLogger(__name__)
@@ -75,337 +69,192 @@ class OutputFileWriter:
                  options: Options,
                  bam_header: dict = None):
 
-        # set the booleans
-        self.write_fastq = options.produce_fastq
-        self.write_bam = options.produce_bam
-        self.write_vcf = options.produce_vcf
-        self.paired = options.paired_ended
-        self.temporary_dir = options.temp_dir_path
-
+        self.paired_ended = options.paired_ended
         self.bam_header = bam_header
 
-        # Set the file names
-
-        self.fastq_fns = None
-        self.fastq1_fn = None
-        self.fastq2_fn = None
-
-        self.bam_fn = None
-        self.vcf_fn = None
+        file_handles = {}
 
         # Set up filenames based on booleans
-        files_to_write = []
-        if self.paired and self.write_fastq:
-            self.fastq1_fn = options.fq1
-            self.fastq2_fn = options.fq2
-            self.fastq_fns = [self.fastq1_fn, self.fastq2_fn]
-            files_to_write.extend(self.fastq_fns)
-        elif self.write_fastq:
-            self.fastq1_fn = options.fq1
-            self.fastq2_fn = self.temporary_dir / "dummy.fastq.gz"
-            self.fastq_fns = [self.fastq1_fn, self.fastq2_fn]
-            files_to_write.extend(self.fastq_fns)
-        if self.write_bam:
-            self.bam_fn = options.bam
-            self.bam_keys = list(bam_header)
-            files_to_write.append(self.bam_fn)
-        if self.write_vcf:
-            self.vcf_fn = options.vcf
-            files_to_write.append(self.vcf_fn)
+        if options.fq1 is not None:
+            fq1 = options.fq1
+            file_handles[fq1] = bgzf.open(fq1, 'wt')
+        else:
+            fq1 = None
+        if options.fq2 is not None:
+            fq2 = options.fq2
+            file_handles[fq2] = bgzf.open(fq2, 'wt')
+        else:
+            fq2 = None
+        if options.vcf is not None:
+            vcf = options.vcf
+            file_handles[vcf] = bgzf.open(vcf, 'wt')
+        else:
+            vcf = None
+        if options.bam is not None:
+            bam = options.bam
+            file_handles[bam] = bgzf.BgzfWriter(bam, 'w', compresslevel=BAM_COMPRESSION_LEVEL)
+        else:
+            bam = None
 
-        self.files_to_write = files_to_write
+        self.fq1 = fq1
+        self.fq2 = fq2
+        self.vcf = vcf
+        self.bam = bam
 
-        # Create files as applicable
-        for file in [x for x in self.files_to_write if x.name != "dummy.fastq.gz"]:
+        if not file_handles:
+            _LOG.error("output_file_writer received no files!")
+            raise ValueError
+
+        # Create files as applicable, also just to check for data collisions
+        for file in file_handles.keys():
             validate_output_path(file, True, options.overwrite_output)
 
-        mode = 'xt'
-        if options.overwrite_output:
-            mode = 'wt'
-        # Initialize the vcf and write the header, if applicable
-        if self.write_vcf:
-            # Writing the vcf header.
-            with open_output(self.vcf_fn, mode=mode) as vcf_file:
-                vcf_file.write(f'##fileformat=VCFv4.1\n')
[…]
