+++ neat/read_simulator/single_runner.py
@@ -1,6 +1,8 @@
 """
 Runner for read-simulator in single-ended mode
 """
+from turtledemo.chaos import coosys
+
 from Bio import SeqIO
 import logging
 import pickle
@@ -8,7 +10,6 @@ import gzip
 
 from .utils import parse_input_vcf, parse_beds, OutputFileWriter, \
     generate_variants, generate_reads, Options
-from ..common import validate_output_path
 from ..variants import ContigVariants
 
 from ..models import MutationModel, SequencingErrorModel, FragmentLengthModel, TraditionalQualityModel
@@ -17,268 +18,92 @@ __all__ = ["read_simulator_single"]
 
 _LOG = logging.getLogger(__name__)
 
-def read_simulator_single(options: Options):
+def read_simulator_single(
+        thread_idx: int,
+        local_options: Options,
+        contig_name: str,
+        mut_model: MutationModel,
+        seq_error_model: SequencingErrorModel,
+        qual_score_model: TraditionalQualityModel,
+        fraglen_model: FragmentLengthModel,
+        input_variants: ContigVariants,
+        target_regions: list,
+        discard_regions: list,
+        mutation_rate_regions: list,
+) -> tuple[]:
     """
-    Model preparation
+    inputs:
+    :param thread_idx: index of current thread
+    :param local_options: options for current thread and reference chunk
+    :param contig_name: The original list of contig names.
+    :param mut_model: The mutation model for the run
+    :param seq_error_model: Sequencing error model for run
+    :param qual_score_model: Quality score model for run
+    :param fraglen_model: Fragment length model for the run
+    :param input_variants: Input variants for the current block
+    :param target_regions: Regions to target in analysis
+    :param discard_regions: Regions to discard from analysis
+    :param mutation_rate_regions: Mutation rate bed regions for this section
+
+    Ideally this should work for either a file chunk or contig. We'll assume here that we're
+    getting either an entire contig or a file chunk, and that no new subdivisions are needed.
+    We can just read in the file.
 
     Read input models or default models, as specified by user.
     """
-
-    _LOG.info(f"Reading Models...")
-
-    (
-        mut_model,
-        seq_error_model_1,
-        seq_error_model_2,
-        qual_score_model_1,
-        qual_score_model_2,
-        fraglen_model
-    ) = initialize_all_models(options)
-
+    if thread_idx != 1:
+        _THREAD_LOG = logging.getLogger(f"thread_{thread_idx}")
+        _THREAD_LOG.propagate = False
+    else:
+        _THREAD_LOG = _LOG
     """
     Process Inputs
     """
-    _LOG.info(f'Reading {options.reference}.')
+    _THREAD_LOG.info(f'Reading {local_options.reference}.')
+    local_ref_index = SeqIO.index(str(local_options.reference), "fasta")
 
-    # TODO check into SeqIO.index_db()
-    reference_index = SeqIO.index(str(options.reference), "fasta")
-    reference_keys_with_lens = {key: len(value) for key, value in reference_index.items()}
-    _LOG.debug("Reference file indexed.")
-
-    if _LOG.getEffectiveLevel() < 20:
-        count = 0
-        for contig in reference_keys_with_lens:
-            count += reference_keys_with_lens[contig]
-        _LOG.debug(f"Length of reference: {count / 1_000_000:.2f} Mb")
-
-    input_variants_dict = {x: ContigVariants() for x in reference_keys_with_lens}
-    if options.include_vcf:
-        _LOG.info(f"Reading input VCF: {options.include_vcf}.")
-        # TODO Check if we need full ref index or just keys and lens
-        sample_names = parse_input_vcf(
-            input_variants_dict,
-            options.include_vcf,
[…]

[…]
