+++ neat/read_simulator/utils/stitch_outputs.py
@@ -16,102 +16,51 @@ import logging
 
 __all__ = ["main"]
 
-from Bio import SeqIO
+from Bio import SeqIO, bgzf
 
 from neat.common import open_output, open_input
-from neat.read_simulator.utils import Options
+from neat.read_simulator.utils import Options, OutputFileWriter
 
 _LOG = logging.getLogger(__name__)
 
-def concat(files_to_join: List[Path], dest: Path) -> None:
+def concat(files_to_join: List[Path], ofw: OutputFileWriter, file: Path) -> None:
     if not files_to_join:
         # Nothing to do, and no error to throw
         return
 
-    with open_output(dest, "wb") as out_f:
-        for f in files_to_join:
-            with open_input(f) as in_f:
-                shutil.copyfileobj(in_f.encode('utf-8'), out_f)
+    out_handle = ofw.files_to_write[file]
+    for f in files_to_join:
+        with bgzf.BgzfReader(f) as in_f:
+            shutil.copyfileobj(in_f, out_handle)
 
-def merge_bam(bams: List[Path], dest: Path) -> None:
+def merge_bam(bams: List[Path], bam_handle) -> None:
     if not bams:
         return
 
-    import pysam
-    unsorted = dest.with_suffix(".unsorted.bam")
-    pysam.merge("--no-PG", "-f", str(unsorted), *map(str, bams))
-    pysam.sort("-o", str(dest), str(unsorted))
-    unsorted.unlink(missing_ok=True)
-    """Make BAM header deterministic: drop @PG, sort @SQ/@RG."""
-    hdr_text = pysam.view("-H", str(dest), text=True)
+    for bam in bams:
+        with bgzf.BgzfReader(bam) as in_bam:
+            for line in in_bam:
+                if line.startswith('@'):
+                    continue
+                else:
+                    bam_handle.write(line)
 
-    hd: list[str] = []
-    sq: list[str] = []
-    rg: list[str] = []
-    other: list[str] = []
-    for line in hdr_text.splitlines():
-        if line.startswith("@PG"):
-            continue
-        if line.startswith("@HD"):
-            hd.append(line)
-        elif line.startswith("@SQ"):
-            sq.append(line)
-        elif line.startswith("@RG"):
-            parts = [t for t in line.split("\t") if not t.startswith("CL:")]
-            line = "\t".join(parts)
-            rg.append(line)
-        else:
-            other.append(line)
-
-    def sn(x: str) -> str:
-        for t in x.split("\t"):
-            if t.startswith("SN:"):
-                return t[3:]
-        return x
-
-    def rgid(x: str) -> str:
-        for t in x.split("\t"):
-            if t.startswith("ID:"):
-                return t[3:]
-        return x
-
-    sq.sort(key=sn)
-    rg.sort(key=rgid)
-
-    new_hdr = "\n".join([
-        *(hd or ["@HD\tVN:1.6\tSO:coordinate"]),
-        *sq,
-        *rg,
-        *other,
-    ]) + "\n"
-
-    hdr_tmp = dest.with_suffix(".reheader.sam")
-    out_tmp = dest.with_suffix(".reheader.bam")
-    hdr_tmp.write_text(new_hdr)
-
-    with open(out_tmp, 'wb') as out_f:
-        data = pysam.reheader(str(hdr_tmp), str(dest))
-        out_f.write(data)
-
-    out_tmp.replace(dest)
-    hdr_tmp.unlink(missing_ok=True)
-
-def main(main_options: Options, thread_tuple: list[tuple[int, Options]]) -> None:
+def main(ofw: OutputFileWriter, output_files: list[tuple[int, dict[str, Path]]]) -> None:
     fq1_list = []
     fq2_list = []
     bam_list = []
[…]
