+++ neat/read_simulator/utils/generate_variants.py
@@ -11,12 +11,12 @@ import re
 import sys
 
 from Bio.Seq import Seq
+from Bio.SeqRecord import SeqRecord
 from numpy.random import Generator
 
-from . import OutputFileWriter
 from ...models import MutationModel
 from ...variants import Insertion, Deletion, SingleNucleotideVariant, ContigVariants
-from ..utils import Options
+from ..utils import Options, bed_func
 from ...common import ALLOWED_NUCL, pick_ploids
 
 _LOG = logging.getLogger(__name__)
@@ -54,7 +54,8 @@ def map_non_n_regions(sequence) -> np.ndarray:
 
 
 def generate_variants(
-        reference: Seq,
+        reference: SeqRecord,
+        ref_start: int,
         mutation_rate_regions: list[tuple[int, int, float]],
         existing_variants: ContigVariants,
         mutation_model: MutationModel,
@@ -64,6 +65,7 @@ def generate_variants(
     """
     This function will generate variants to add to the dataset, by writing them to the input temp vcf file.
     :param reference: The reference to generate variants off of. This should be a SeqRecord object
+    :param ref_start: where on the reference this SeqRecord above starts.
     :param mutation_rate_regions: Genome segments with associated mutation rates
     :param existing_variants: Any input variants or overlaps to pick up from the previous contig
     :param mutation_model: The mutation model for the dataset
@@ -80,7 +82,7 @@ def generate_variants(
     max_window_size = 1000
     n_added = 0
 
-    start = time.time()
+    start_time = time.time()
 
     """
     Decide how many and where any random mutations will happen in this contig
@@ -97,7 +99,8 @@ def generate_variants(
     factors.append(total * mutation_model.avg_mut_rate)
     overall_mutation_average = sum(factors)/len(reference)
     average_number_of_mutations = round(len(reference) * overall_mutation_average)
-    max_mutations = round(len(reference) * 0.3)
+    # Trying to set some sort of rational upper bound to restrain the poisson distribution which, rarely, blows up.
+    max_mutations = round(len(reference) * (overall_mutation_average * 2))
 
     # This number will serve as a counter for our loops. Default is 1 mutation per chromosome.
     min_mutations = options.min_mutations
@@ -106,27 +109,32 @@ def generate_variants(
     # Pick a random number from a poisson distribution. We want at least min_mutations and at most max mutations.
     how_many_mutations = min(max(options.rng.poisson(average_number_of_mutations), min_mutations), max_mutations)
 
-    _LOG.debug(f'Planning to add {how_many_mutations} mutations. The final number may be less.')
+    _LOG.info(f'Planning to add {how_many_mutations} mutations. The final number may be less.')
 
     while how_many_mutations > 0:
         # Pick a region based on the mutation rates
         # (default is one rate for the whole chromosome, so this will be trivial in that case
         # for this selection, we'll normalize the mutation rates
         probability_rates = mutation_rates / sum(mutation_rates)
-        mut_region = options.rng.choice(a=mutation_rate_regions, p=probability_rates)
+        # We need to intersect our chosen mutation region with our block
+        local_mut_regions = bed_func.intersect_regions(mutation_rate_regions, (ref_start, ref_start + len(reference)), options.mutation_rate)
+        # For no input mutation regions bed, this will return the entire sequence.
+        mut_region = options.rng.choice(a=local_mut_regions, p=probability_rates)
+        mut_region_offset = (int(mut_region[0]-ref_start), int(mut_region[1]-ref_start), mut_region[2])
         # Pick a random starting place. Randint is inclusive of endpoints, so we subtract 1
[…]
