### 2025-10-12 8ab3b88
Working but something is slowing us way down. Unclear from the graph what that is
