### 2025-10-13 a3bba2f
It's working, but writing the bam painfully slowly for reasons I don't understand
