### 2025-10-13 572adff
Almost everything is working, but bam creation is significantly slower than it used to be
