### 2025-10-13 44bb242
Still tweaking bam creation. Next thing I'll try is creating the smaller bams first then stitching them together, like we were doing before.
