### 2025-10-14 cdb1abb
Got bam production up and running again
