+++ neat/read_simulator/utils/stitch_outputs.py
@@ -7,6 +7,7 @@ import gzip
 import pickle
 import re
 import shutil
+import pysam
 import subprocess
 import sys
 from multiprocessing import Pool, Process
@@ -23,7 +24,6 @@ from Bio import SeqIO, bgzf
 
 from neat.common import open_output, open_input
 from neat.read_simulator.utils import Options, OutputFileWriter
-from neat.read_simulator.utils.output_file_writer import BAM_COMPRESSION_LEVEL
 
 _LOG = logging.getLogger(__name__)
 
@@ -36,55 +36,37 @@ def concat(files_to_join: List[Path], ofw: OutputFileWriter, file: Path) -> None
     for f in files_to_join:
         with bgzf.BgzfReader(f) as in_f:
             shutil.copyfileobj(in_f, out_handle)
-    out_handle.close()
 
-def merge_bam(reads_pickles: List[Path], ofw: OutputFileWriter, contig_dict: dict, read_length: int, threads: int) -> None:
-    if not reads_pickles:
+def merge_bam(bam_files: List[Path], ofw: OutputFileWriter, threads: int) -> None:
+    if not bam_files:
         return
 
-    bam_handle = bgzf.BgzfWriter(ofw.bam, "w", compresslevel=BAM_COMPRESSION_LEVEL)
-    for file in reads_pickles:
-        contig_reads_data = pickle.load(gzip.open(file))
-        for read_data in contig_reads_data:
-            read1 = read_data[0]
-            read2 = read_data[1]
-            if read1:
-                ofw.write_bam_record(
-                    read1,
-                    contig_dict[read1.reference_id],
-                    bam_handle,
-                    read_length
-                )
-            if read2:
-                ofw.write_bam_record(
-                    read2,
-                    contig_dict[read2.reference_id],
-                    bam_handle,
-                    read_length
-                )
+    unsorted = ofw.bam.with_suffix(".unsorted.bam")
+    sorted_bam_files = []
+    pysam.merge("--no-PG", "-@", str(threads), "-f", str(unsorted), *map(str, bam_files))
+    pysam.sort("-@", str(threads), "-o", str(ofw.bam), str(unsorted))
+    unsorted.unlink(missing_ok=True)
 
 def main(
         ofw: OutputFileWriter,
         output_files: list[tuple[int, str, dict[str, Path]]],
-        contig_dict: dict | None = None,
-        read_length: int | None = None,
         threads: int | None = None
 ) -> None:
 
     fq1_list = []
     fq2_list = []
-    reads_pickles = []
+    bam = []
     # Gather all output files from the ops objects
     for (thread_idx,file_dict) in output_files:
         if file_dict["fq1"]:
             fq1_list.append(file_dict["fq1"])
         if file_dict["fq2"]:
             fq2_list.append(file_dict["fq2"])
-        if file_dict["reads"]:
-            reads_pickles.append(file_dict["reads"])
+        if file_dict["bam"]:
+            bam.append(file_dict["bam"])
     # concatenate all files of each type. An empty list will result in no action
     concat(fq1_list, ofw, ofw.fq1)
     concat(fq2_list, ofw, ofw.fq2)
-    merge_bam(reads_pickles, ofw, contig_dict, read_length, threads)
+    merge_bam(bam, ofw, threads)
     # Final success message via logging
     _LOG.info("Stitching complete!")

[…]
