+++ neat/read_simulator/utils/stitch_outputs.py
@@ -21,28 +21,27 @@ import logging
 __all__ = ["main"]
 
 from Bio import SeqIO, bgzf
+from Bio.bgzf import BgzfWriter
 
 from neat.common import open_output, open_input
 from neat.read_simulator.utils import Options, OutputFileWriter
 
 _LOG = logging.getLogger(__name__)
 
-def concat(files_to_join: List[Path], ofw: OutputFileWriter, file: Path) -> None:
+def concat(files_to_join: List[Path], dest_file: BgzfWriter) -> None:
     if not files_to_join:
         # Nothing to do, and no error to throw
         return
 
-    out_handle = ofw.files_to_write[file]
     for f in files_to_join:
         with bgzf.BgzfReader(f) as in_f:
-            shutil.copyfileobj(in_f, out_handle)
+            shutil.copyfileobj(in_f, dest_file)
 
 def merge_bam(bam_files: List[Path], ofw: OutputFileWriter, threads: int) -> None:
     if not bam_files:
         return
 
     unsorted = ofw.bam.with_suffix(".unsorted.bam")
-    sorted_bam_files = []
     pysam.merge("--no-PG", "-@", str(threads), "-f", str(unsorted), *map(str, bam_files))
     pysam.sort("-@", str(threads), "-o", str(ofw.bam), str(unsorted))
     unsorted.unlink(missing_ok=True)
@@ -65,8 +64,8 @@ def main(
         if file_dict["bam"]:
             bam.append(file_dict["bam"])
     # concatenate all files of each type. An empty list will result in no action
-    concat(fq1_list, ofw, ofw.fq1)
-    concat(fq2_list, ofw, ofw.fq2)
+    concat(fq1_list, ofw.files_to_write[ofw.fq1])
+    concat(fq2_list, ofw.files_to_write[ofw.fq2])
     merge_bam(bam, ofw, threads)
     # Final success message via logging
     _LOG.info("Stitching complete!")
