### 2025-10-14 40e4f25
slight adjustment to python-app for  github
