+++ .github/workflows/python-app.yml
@@ -2,7 +2,7 @@ name: NEAT Unit Tests
 
 on:
   push:
-    branches: [develop, main]
+    branches: [develop]
   pull_request:
     branches: [main]
 
