### 2025-10-14 76027fb
slight adjustment to python-app for  github
