+++ .github/workflows/python-app.yml
@@ -1,8 +1,6 @@
 name: NEAT Unit Tests
 
 on:
-  push:
-    branches: [develop]
   pull_request:
     branches: [main]
 
