### 2025-10-15 cbd114f
Bam creation working again
