### 2025-10-15 f818c75
Updated some logs
