### 2025-10-15 4514a12
Fixed a bug in produce_bam = false
