+++ BacterialWrapperScript/BacterialWrapperScript.py
@@ -1,4 +1,14 @@
 import subprocess
+import gzip
+import shutil
+import yaml
+import pysam
+
+from pathlib import Path
+from typing import List
+from neat.read_simulator.utils import Options, OutputFileWriter
+from Bio import bgzf
+from Bio.bgzf import BgzfWriter
 
 
 # Rearranges the bacterial chromosome by wrapping it around
@@ -101,9 +111,185 @@ salmonella_config_file = "Outputs/Salmonella/salmonella_config_test.yml"
 pneumonia_config_file = "Outputs/Pneumonia/pneumonia_config_test.yml"
 tuberculosis_config_file = "Outputs/Tuberculosis/tuberculosis_config_test.yml"
 
-
 # Call the bacterial wrapper script on each reference file
 
-bacterial_wrapper(salmonella_file, "Salmonella", salmonella_config_file)
-bacterial_wrapper(pneumonia_file, "Pneumonia", pneumonia_config_file)
-bacterial_wrapper(tuberculosis_file, "Tuberculosis", tuberculosis_config_file)
+# bacterial_wrapper(salmonella_file, "Salmonella", salmonella_config_file)
+# bacterial_wrapper(pneumonia_file, "Pneumonia", pneumonia_config_file)
+# bacterial_wrapper(tuberculosis_file, "Tuberculosis", tuberculosis_config_file)
+
+
+
+# Stitching outputs back together - based on Keshav's script
+
+# def stitch_outputs(input_files, bacteria_name):
+#     dest_file = Path(bacteria_name + "_stitched_reads.fastq")
+#     with dest_file.open("wb") as out_f:
+#         for input_file in input_files:
+#             input_file = Path(input_file)
+            
+#             if input_file.suffix in {".gz", "bgz"}:
+#                 opener = gzip.open
+#             else:
+#                 opener = open
+#             with opener(input_file, "rb") as in_f:
+#                 shutil.copyfileobj(in_f, out_f)
+
+
+# Stitch output fastq files together for each reference file
+
+# orig_salmonella_input = "Outputs/Salmonella/salmonella_config_test.fastq"
+# new_salmonella_input = "Outputs/Salmonella/new_Salmonella_config_test.fastq"
+# salmonella_inputs = [orig_salmonella_input, new_salmonella_input]
+# stitch_outputs(salmonella_inputs, "salmonella")
+
+# orig_pneumonia_input = "Outputs/Pneumonia/pneumonia_config_test.fastq"
+# new_pneumonia_input = "Outputs/Pneumonia/new_Pneumonia_config_test.fastq"
+# pneumonia_inputs = [orig_pneumonia_input, new_pneumonia_input]
+# stitch_outputs(pneumonia_inputs, "pneumonia")
+
+# orig_tuberculosis_input = "Outputs/Tuberculosis/tuberculosis_config_test.fastq"
+# new_tuberculosis_input = "Outputs/Tuberculosis/new_Tuberculosis_config_test.fastq"
+# tuberculosis_inputs = [orig_tuberculosis_input, new_tuberculosis_input]
+# stitch_outputs(tuberculosis_inputs, "tuberculosis")
+
+
+
+
+
+
+# Stitching all outputs together - Keshav's script
+
+def concat_fq(input_files: List[Path], dest: BgzfWriter) -> None:
+    
+    if not input_files:
+        return
+    
+    for input_file in input_files:
+        with bgzf.BgzfReader(input_file) as in_f:
+            shutil.copyfileobj(in_f, dest)
+
+# def concat_fq(input_files: List[Path], dest: Path) -> None:
+    
+#     if not input_files:
+#         return
+    
[…]
