### 2025-10-16 edc67b8
Worked on testing functions for stitching outputs together
