### 2025-10-16 17a5340
Removed reference and data files
