### 2025-10-16 541b3b1
Removed reference and data files
