### 2025-10-17 a1ef7bb
Vcf seems to be sorted out and it is running much faster in parallel mode now.
