### 2025-10-17 d42dd76
Updating changelog and readme. I missed a couple releases.
