### 2025-10-23 dde0d28
Fixing quality score array output
