### 2025-10-24 a22f72a
added functions to index the bam and vcf files
