+++ neat/read_simulator/utils/output_file_writer.py
@@ -14,6 +14,7 @@ from struct import pack
 import logging
 from typing import Any
 
+import pysam
 from Bio import bgzf
 from pathlib import Path
 
@@ -114,10 +115,11 @@ class OutputFileWriter:
             # Writing the vcf header.
             vcf_header = f'##fileformat=VCFv4.1\n' \
                          f'##reference={Path(options.reference).resolve()}\n' \
-                         f'##Generated by NEAT with RNG value: {options.rng_seed}\n' \
+                         f'##source=NEAT\n' \
+                         f'##RNG_seed={options.rng_seed}\n' \
                          f'##ALT=<ID=DEL,Description="Deletion">\n' \
                          f'##ALT=<ID=INS,Description="Insertion of novel sequence">\n' \
-                         f'##ALT=<ID=SNP,Description="Single Nucleotide Polymorhism">\n' \
+                         f'##ALT=<ID=SNP,Description="Single Nucleotide Polymorphism">\n' \
                          f'##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">\n' \
                          f'#CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO\tFORMAT\tNEAT_simulated_sample\n'
             self.files_to_write[self.vcf].write(vcf_header)
@@ -150,7 +152,7 @@ class OutputFileWriter:
             _LOG.error(f"Tried to write fastq record to unknown file {filename}")
             raise ValueError
 
-    def flush_and_close_files(self, no_bam: bool = False):
+    def flush_and_close_files(self, no_bam: bool = True, index_file: bool = False):
         for file_name in self.files_to_write:
             if file_name.suffix == "bam" and no_bam:
                 continue
