### 2025-10-24 71fc59a
Added some clarifying comments
