### 2025-11-01 469500a
Updates to constrain read creation
