### 2025-11-02 a732726
bug in stitch fixed
