+++ neat/read_simulator/utils/stitch_outputs.py
@@ -2,37 +2,21 @@
 Stitch NEAT split‑run outputs into one dataset.
 """
 
-import argparse
-import gzip
-import pickle
-import re
 import shutil
 import pysam
-import subprocess
-import sys
-from multiprocessing import Pool, Process
 from pathlib import Path
-from struct import pack
-from typing import Iterable, List, Tuple
-import yaml
+from typing import List
 
 import logging
 
 __all__ = ["main"]
 
-from Bio import SeqIO, bgzf
-from Bio.bgzf import BgzfWriter
-
-from neat.common import open_output, open_input
-from neat.read_simulator.utils import Options, OutputFileWriter
+from Bio import bgzf
+from neat.read_simulator.utils import OutputFileWriter
 
 _LOG = logging.getLogger(__name__)
 
 def concat(files_to_join: List[Path], ofw: OutputFileWriter, read_num: str) -> None:
-    if not files_to_join:
-        # Nothing to do, and no error to throw
-        return
-
     read_file = ofw.fq1 if read_num == "1" else ofw.fq2
     dest_file = ofw.files_to_write[read_file]
 
@@ -40,9 +24,8 @@ def concat(files_to_join: List[Path], ofw: OutputFileWriter, read_num: str) -> N
         with bgzf.BgzfReader(f) as in_f:
             shutil.copyfileobj(in_f, dest_file)
 
-def merge_vcfs(vcfs: List[Path], dest: BgzfWriter) -> None:
-    if not vcfs:
-        return
+def merge_vcfs(vcfs: List[Path], ofw: OutputFileWriter) -> None:
+    dest = ofw.files_to_write[ofw.vcf]
     for vcf in vcfs:
         with bgzf.BgzfReader(vcf) as fh:
             for line in fh:
@@ -50,9 +33,6 @@ def merge_vcfs(vcfs: List[Path], dest: BgzfWriter) -> None:
                     dest.write(line)
 
 def merge_bam(bam_files: List[Path], ofw: OutputFileWriter, threads: int) -> None:
-    if not bam_files:
-        return
-
     # ofw.files_to_write[ofw.bam].close()
     unsorted = ofw.bam.with_suffix(".unsorted.bam")
     pysam.merge("--no-PG", "-@", str(threads), "-f", str(unsorted), *map(str, bam_files))
