### 2025-11-04 5a8bb33
Fixing bugs and code cleanup. Vcf looking better.
