+++ neat/read_simulator/utils/generate_variants.py
@@ -52,12 +52,11 @@ def map_non_n_regions(sequence) -> np.ndarray:
 
     return base_check
 
-# TODO tests to ensure generate variants doesn't try to sample out of bounds
 def generate_variants(
         reference: SeqRecord,
         ref_start: int,
         mutation_rate_regions: list[tuple[int, int, float]],
-        existing_variants: ContigVariants,
+        input_variants: ContigVariants,
         mutation_model: MutationModel,
         options: Options,
         max_qual_score: int,
@@ -67,13 +66,16 @@ def generate_variants(
     :param reference: The reference to generate variants off of. This should be a SeqRecord object
     :param ref_start: where on the reference this SeqRecord above starts.
     :param mutation_rate_regions: Genome segments with associated mutation rates
-    :param existing_variants: Any input variants or overlaps to pick up from the previous contig
+    :param input_variants: Any input variants or overlaps to pick up from the previous contig
     :param mutation_model: The mutation model for the dataset
     :param options: The options for the run
     :param max_qual_score: the maximum quality score for the run.
     """
-    return_variants = existing_variants
-    # Step 1: Create a VCF of mutations
+    return_variants = ContigVariants()
+    for variant_location in input_variants.variant_locations:
+        if ref_start <= variant_location < ref_start + len(reference):
+            for variant in input_variants.contig_variants[variant_location]:
+                return_variants.add_variant(variant)
 
     # pase out the mutation rates
     mutation_rates = np.array([x[2] for x in mutation_rate_regions])
@@ -124,30 +126,29 @@ def generate_variants(
 
         # Pick a random starting place. Randint is inclusive of endpoints, so we subtract 1
         if mut_region_offset[1] <= mut_region_offset[0]:
-            # ensure we are still in bounds
-            continue
+            _LOG.error("Invalid Mutation Region")
+            raise ValueError
 
         window_start = options.rng.integers(mut_region_offset[0], mut_region_offset[1] - 1, dtype=int)
         found = False
+        if window_start > len(reference):
+            _LOG.error("Invalid Mutation Region")
+            raise ValueError
         if reference[window_start] not in ALLOWED_NUCL:
-            # pick a random location to the right
-            if window_start + 1 >= mut_region_offset[1] - 1 or mut_region_offset[1] -1 <= 0:
-                # ensure we are still in bounds
-                continue
-            plus = options.rng.integers(window_start + 1, mut_region_offset[1] - 1, dtype=int)
-            if reference[plus] in ALLOWED_NUCL:
-                found = True
-                window_start = plus
-            else:
-                # If that didn't work pick a random location to the left
-                if mut_region_offset[0] >= window_start - 1 or window_start - 1 <= 0:
-                    # ensure we are still in bounds
-                    continue
-
-                minus = options.rng.integers(mut_region_offset[0], window_start - 1, dtype=int)
-                if reference[minus] in ALLOWED_NUCL:
+            # Ensure we are not too close to the edge
+            if window_start + 1 < mut_region_offset[1] - 1 and mut_region_offset[1] - 1 > 0:
+                # pick a random location to the right
+                plus = options.rng.integers(window_start + 1, mut_region_offset[1] - 1, dtype=int)
+                if reference[plus] in ALLOWED_NUCL:
                     found = True
-                    window_start = minus
+                    window_start = plus
+                else:
+                    # If that didn't work pick a random location to the left
+                    if mut_region_offset[0] < window_start - 1 and window_start - 1 > 0:
+                        minus = options.rng.integers(mut_region_offset[0], window_start - 1, dtype=int)
+                        if reference[minus] in ALLOWED_NUCL:
+                            found = True
+                            window_start = minus
         else:
[…]
