+++ neat/read_simulator/runner.py
@@ -2,6 +2,7 @@
 Runner for generate_reads task
 """
 import logging
+import os
 import time
 import multiprocessing as mp
 
@@ -130,7 +131,8 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
     output_opts: list = []
     output_files: list = []
     # a dict with contig keys, then the thread index, and finally the applicable contig variants as the value
-    all_variants: dict[str, dict[int, ContigVariants]] = {chrom: {} for chrom in reference_index.keys()}
+    # TODO Remove if not needed
+    # all_variants: dict[str, dict[int, ContigVariants]] = {chrom: {} for chrom in reference_index.keys()}
     thread_idx = 1
     contig_list = list(reference_keys_with_lens.keys())
     contig_dict = {contig: contig_list.index(contig) for contig in reference_keys_with_lens.keys()}
@@ -173,7 +175,8 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
                     mutation_rate_dict[contig],
                 )
                 _LOG.info(f"Completed simulating contig {contig}.")
-                all_variants[contig][idx] = local_variants
+                # TODO Remove if not needed
+                # all_variants[contig][idx] = local_variants
                 output_files.append((thread_idx, files_written))
             else:
                 thread_input_variants = filter_thread_variants(input_variants_dict[contig], (start, start+length))
@@ -205,7 +208,8 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
 
         # Need to organize the results, as above
         for idx, contig, local_variants, files_written in results.get():
-            all_variants[contig][idx] = local_variants
+            # TODO Remove if not needed
+            # all_variants[contig][idx] = local_variants
             output_files.append((idx, files_written))
 
     _LOG.info("Processing complete, writing output")
@@ -229,7 +233,6 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
         elif "fastq" in file.name:
             continue
         else:
-            # Must be a vcf
             pysam.tabix_index(str(file), preset="vcf", force=force)
 
     _LOG.info(f"Read simulator complete in {time.time() - analysis_start} s")
