### 2025-11-04 5a5f54d
fixing some issues
