### 2025-11-04 73ebf49
Fixed vcf creation
