+++ neat/gen_mut_model/runner.py
@@ -49,7 +49,7 @@ def runner(reference_index,
     :param Path outcounts_file: Path to outcounts file, if present
     :param bool show_trinuc: If true, display trinucleotide counts
     :param bool save_trinuc: If true, save trinucleotide counts
-    :param Path output: Full path to final output
+    :param Path output: File to write
     :param Path bed: Optional input bed
     :param bool human_sample: If true, treat as human sample
     :param bool skip_common: if true, skip common
@@ -341,7 +341,8 @@ def compute_mut_runner(reference,
                        save_trinuc,
                        human_sample,
                        skip_common,
-                       output,
+                       output_dir,
+                       output_prefix,
                        overwrite_output):
     """
     Runner for computing the mutation model
@@ -354,13 +355,17 @@ def compute_mut_runner(reference,
     :param bool save_trinuc: Optionally save trinucleotide counts
     :param bool human_sample: If true, treat as human sample
     :param bool skip_common: if true, skip common variants
-    :param str output: path to output file
+    :param str output_dir: directory to write output files
+    :param str output_prefix: File prefix for output
     :param bool overwrite_output: True to overwrite output
     """
 
     validate_input_path(reference)
     validate_input_path(mutations)
     mutations = Path(mutations)
+    validate_output_path(output_dir, is_file=False)
+    outcounts_prefix = output_prefix + ".trinuc.pickle.gz"
+    model_prefix = output_prefix + ".model.pickle.gz"
 
     if bed:
         validate_input_path(bed)
@@ -370,7 +375,7 @@ def compute_mut_runner(reference,
         validate_input_path(outcounts)
         outcounts = Path(outcounts)
     elif save_trinuc:
-        outcounts = Path(output + '.trinuc.pickle.gz')
+        outcounts = Path(output_dir) / outcounts_prefix
 
     print('Processing reference...')
     reference_index = SeqIO.index(reference, 'fasta')
@@ -440,7 +445,7 @@ def compute_mut_runner(reference,
     else:
         vcf_to_process = mutations
 
-    output = Path(output + '.pickle.gz')
+    output = Path(output_dir) / model_prefix
     validate_output_path(output, overwrite=overwrite_output)
 
     runner(
