### 2025-11-04 0331810
Adding scipy to pyproject.toml
