+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "neat"
-version = "4.3.2"
+version = "4.3.5"
 description = "NGS Simulation toolkit"
 authors = ["Joshua Allen <jallen17@illinois.edu>"]
 license = "BSD 3-Clause License"
