### 2025-11-05 919be52
simplifying the installation hopefully
