+++ README.md
@@ -58,7 +58,7 @@ use the poetry module in build a wheel file, which can then be pip installed. Yo
 commands from within the NEAT directory.
 
 ```
-$ conda env create -f environment.yml -n neat
+$ conda env create -f environment.yml
 $ conda activate neat
 $ poetry build
 $ pip install dist/neat*whl
@@ -70,7 +70,7 @@ This allows you to run NEAT as a command line tool directly:
 Alternatively, if you wish to work with NEAT in the development environment, you can use poetry install within
 the NEAT repo, after creating the conda environment:
 ```
-$ conda env create -f environment.yml -n neat
+$ conda env create -f environment.yml
 $ conda activate neat
 $ poetry install
 ```
