### 2025-11-05 35b3894
modeling scripts work againg
