+++ README.md
@@ -134,7 +134,7 @@ The default is given:
 `mutation_model`: full path to a mutation model generated by NEAT. Leave empty to use a default model _(default model based on human data sequenced by Illumina)_     
 `fragment_model`: full path to fragment length model generate by NEAT. Leave empty to use default model _(default model based on human data sequenced by Illumina)_
 
-`threads`: The number of threads for NEAT to use. _Note: this feature is not yet fully implemented_
+`threads`: The number of threads for NEAT to use. Default is 1.
 `avg_seq_error`: average sequencing error rate for the sequencing machine. Use to increase or decrease the rate of errors in the reads. _Float between 0 and 0.3. Default is set by the error model._    
 `rescale_qualities`: rescale the quality scores to reflect the avg_seq_error rate above. Set True to activate if you  notice issues with the sequencing error rates in your datatset.    
 `include_vcf`: full path to list of variants in vcf format to include in the simulation. These will be inserted as they appear in the input VCF into the final VCF, and the corresponding fastq and bam files, if requested.    
@@ -338,13 +338,13 @@ neat gen-mut-model reference.fa input_variants.vcf   \
 
 Trinucleotides are identified in the reference genome and the variant file. Frequencies of each trinucleotide transition are calculated and output as a pickle (.p) file.
 
-| Option          | Description                                                                  |
-|-----------------|------------------------------------------------------------------------------|
-| -o <str>        | Path to output file and prefix. Required.                                    |
-| --bed           | Flag that indicates you are using a bed-restricted vcf and fasta (see below) |
-| --save-trinuc   | Save trinucleotide counts for reference                                      |
-| --human-sample  | Use to skip unnumbered scaffolds in human references                         |
-| --skip-common   | Do not save common snps or high mutation areas                               |
+| Option          | Description                                                                   |
+|-----------------|-------------------------------------------------------------------------------|
+| -o <str>        | Path to output file and prefix. Defaults to "neat_sim" in current working dir |
+| --bed           | Flag that indicates you are using a bed-restricted vcf and fasta (see below)  |
+| --save-trinuc   | Save trinucleotide counts for reference                                       |
+| --human-sample  | Use to skip unnumbered scaffolds in human references                          |
+| --skip-common   | Do not save common snps or high mutation areas                                |
 
 ## neat model-seq-err
 
