### 2025-11-06 #9056 @keshav-gandhi
Hello @csoneson, thanks for reaching out! Some users of our tool have found new bugs with the release of NEAT v4.3.2. Josh and I are currently working on fixing the bugs/testing and are in the process of rolling out a cleaner version of the software.
