### 2025-11-07 c2ab343
Refactor import statement for pairwise2 module
Unneeded submodule
