### 2025-11-07 b0ddef4
Fix for model_seq_error
