### 2025-11-07 d3e656b
added an exit if file can't be read correctly
