### 2025-11-07 5c6b4ac
Adding log message
