### 2025-11-07 97e1b92
Removed 'upper()' calls when writing, since we did 'upper()' when reading them in.
