### 2025-11-07 da0b030
fixed fastq qual scores
