### 2025-11-07 4608920
fixing bam qual scores
