### 2025-11-07 e1aba21
Fixed bam quality scores
