### 2025-11-07 f27cb86
Hoping to speed up intermediate file I/O
