### 2025-11-07 0ae9c86
Hoping to speed up intermediate file I/O
