### 2025-11-09 d0089d5
Bam creation working as expected again
