+++ neat/read_simulator/utils/stitch_outputs.py
@@ -17,22 +17,18 @@ from neat.read_simulator.utils import OutputFileWriter
 
 _LOG = logging.getLogger(__name__)
 
-def concat(files_to_join: List[Path], file_to_write: Path) -> None:
-    dest_file = bgzf.BgzfWriter(file_to_write)
+def concat(files_to_join: List[Path], dest_file: gzip.GzipFile) -> None:
     for f in files_to_join:
-        with gzip.open(f, 'r') as in_f:
-            for line in in_f:
-                dest_file.write(line)
-    dest_file.close()
+        with gzip.open(f, 'rt') as in_f:
+            shutil.copyfileobj(in_f, dest_file)
 
-def merge_vcfs(vcfs: List[Path], file_to_write: Path) -> None:
-    dest = bgzf.BgzfWriter(file_to_write)
+def merge_vcfs(vcfs: List[Path], ofw: OutputFileWriter) -> None:
+    dest = ofw.files_to_write[ofw.vcf]
     for vcf in vcfs:
         with gzip.open(vcf, 'rt') as fh:
             for line in fh:
                 if not line.startswith("#"):
                     dest.write(line)
-    dest.close()
 
 def merge_bam(bam_files: List[Path], ofw: OutputFileWriter, threads: int) -> None:
     unsorted = ofw.bam.with_suffix(".unsorted.bam")
@@ -62,11 +58,11 @@ def main(
             bam_list.append(file_dict["bam"])
     # concatenate all files of each type. An empty list will result in no action
     if fq1_list:
-        concat(fq1_list, ofw.fq1)
+        concat(fq1_list, ofw.files_to_write[ofw.fq1])
     if fq2_list:
-        concat(fq2_list, ofw.fq2)
+        concat(fq2_list, ofw.files_to_write[ofw.fq2])
     if vcf_list:
-        merge_vcfs(vcf_list, ofw.vcf)
+        merge_vcfs(vcf_list, ofw)
     if bam_list:
         merge_bam(bam_list, ofw, threads)
     # Final success message via logging
