### 2025-11-10 2252667
Modified how vcf header was written and when to get the indexing to work right
