+++ neat/read_simulator/runner.py
@@ -9,6 +9,7 @@ import multiprocessing as mp
 from pathlib import Path
 
 import pysam
+from pysam import bcftools
 from Bio import SeqIO
 
 from .utils import Options, OutputFileWriter, parse_beds, parse_input_vcf
@@ -114,7 +115,7 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
     # Creates files and sets up objects for files that can be written to as needed.
     # Also creates headers for bam and vcf. We create the overall bam with no header, as it will get a header from
     # merging the smaller bams.
-    output_file_writer = OutputFileWriter(options=options, bam_header=None)
+    output_file_writer = OutputFileWriter(options=options, vcf_header = reference_keys_with_lens, bam_header=None)
 
     # Split file by chunk for parallel analysis or by contig for either parallel or single analysis
     _LOG.info("Splitting reference...")
@@ -233,7 +234,7 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
         elif "fastq" in file.name:
             continue
         else:
-            # may still need to sort the vcf
+            bcftools.sort(str(file), "-Ob9")
             pysam.tabix_index(str(file), preset="vcf", force=force)
 
     _LOG.info(f"Read simulator complete in {time.time() - analysis_start} s")
