+++ neat/read_simulator/utils/output_file_writer.py
@@ -66,15 +66,18 @@ class OutputFileWriter:
     in the various formats.
 
     :param options: Options for the current run.
+    :param vcf_header: optional parameters for vcf header.
     :param bam_header: A dictionary of lengths of each contig from the reference, keyed by contig id.
     :param vcf_format: The format to compress the vcf file, since we need to speed up the intermediate files with gzip
     """
     def __init__(self,
                  options: Options,
+                 vcf_header: dict = None,
                  vcf_format: str = "bgzip",
                  bam_header: dict = None):
 
         self.paired_ended = options.paired_ended
+        self.vcf_header = vcf_header
         self.bam_header = bam_header
         self.vcf_format = vcf_format
 
@@ -116,8 +119,11 @@ class OutputFileWriter:
         self.files_to_write = file_handles
 
         # Initialize the vcf and write the header, if applicable
-        if options.produce_vcf:
+        if options.produce_vcf and self.vcf_header:
             # Writing the vcf header.
+            ref_line = ""
+            for (contig, length) in self.vcf_header.items():
+                ref_line += f"##contig=<ID={contig}, length={length}>\n"
             vcf_header = f'##fileformat=VCFv4.1\n' \
                          f'##reference={Path(options.reference).resolve()}\n' \
                          f'##source=NEAT\n' \
@@ -126,6 +132,7 @@ class OutputFileWriter:
                          f'##ALT=<ID=INS,Description="Insertion of novel sequence">\n' \
                          f'##ALT=<ID=SNP,Description="Single Nucleotide Polymorphism">\n' \
                          f'##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">\n' \
+                         f'{ref_line}' \
                          f'#CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO\tFORMAT\tNEAT_simulated_sample\n'
             self.files_to_write[self.vcf].write(vcf_header)
 
