### 2025-11-10 d66d5a9
Fixing sort command, changing to subprocess
