### 2025-11-11 6ea24d4
BAM alignments looking better
