+++ neat/read_simulator/utils/read.py
@@ -8,12 +8,13 @@ Methods allow comparisons between reads, based on chromosome, start and end. Als
 both the reference sequence and the read and the actual read sequence.
 """
 import logging
+from copy import deepcopy
+
 import numpy as np
 
 from typing import TextIO, Iterator
 from Bio.Seq import Seq, MutableSeq
-from Bio import pairwise2
-from Bio.pairwise2 import format_alignment
+from Bio.Align import PairwiseAligner
 from numpy.random import Generator
 
 from ...common import ALLOWED_NUCL
@@ -304,7 +305,7 @@ class Read:
             self,
             err_model: SequencingErrorModel,
             qual_model: TraditionalQualityModel,
-            fastq_handle: Iterator,
+            fastq_handle,
             quality_offset: int,
             produce_fastq: bool,
             rng: Generator,
@@ -314,7 +315,7 @@ class Read:
 
         :param err_model: The error model for the run
         :param qual_model: The quality score model for the run
-        :param fastq_handle: for writing fastq1
+        :param fastq_handle: for writing fastq1, either a gzip file or a bgzipWriter object.
         :param quality_offset: the quality offset for this run
         :param produce_fastq: If true, this will write out the temp fastqs. If false, this will only write out the tsams
             to create the bam files.
@@ -332,8 +333,8 @@ class Read:
         # It updates the quality array and reference segment in place, including reversing them, if appropriate
         self.convert_masking(qual_model)
 
-        # set the read sequence to match the reference
-        self.read_sequence = self.reference_segment
+        # set the read sequence to match the reference, then modify
+        self.read_sequence = deepcopy(self.reference_segment)
 
         # Get errors for the rea and update the quality score
         self.errors, self.padding = err_model.get_sequencing_errors(
@@ -352,9 +353,13 @@ class Read:
         if not self.mapping_quality:
             self.mapping_quality = 70
 
+        if self.is_reverse:
+            self.read_sequence = self.read_sequence.reverse_complement()[:self.run_read_length]
+        else:
+            self.read_sequence = self.read_sequence[:self.run_read_length]
         # Might as well double check here
         if produce_fastq:
-            fastq_record = f'@{self.name}\n{str(self.read_sequence[:self.run_read_length])}\n+\n{self.read_quality_string}\n'
+            fastq_record = f'@{self.name}\n{str(self.read_sequence)}\n+\n{self.read_quality_string}\n'
             fastq_handle.write(fastq_record)
 
 
@@ -369,11 +374,7 @@ class Read:
         bad_score = min(quality_model.quality_scores)
         # we'll use generic human repeats, as commonly found in masked regions. We may refine this to make configurable
         repeat_bases = list("TTAGGG")
-        if self.is_reverse:
-            raw_sequence = self.reference_segment.reverse_complement()
-            self.quality_array = self.quality_array[::-1]
-        else:
-            raw_sequence = self.reference_segment
+        raw_sequence = deepcopy(self.reference_segment)
 
         start = raw_sequence.find('N')
         if start != -1:
@@ -390,64 +391,6 @@ class Read:
 
         self.reference_segment = Seq(modified_segment)
 
-    def align_seqs(self):
-        """
-        The sequence alignment. We restrict the alignment to the section of the reference where we know the read
-        came from and try to generate a minimal cigar string. The cigar string part may still need tweaking.
-        """
-        raw_alignment = pairwise2.align.globalms(
-            self.reference_segment,
-            self.read_sequence,
-            match=10,
-            mismatch=-10,
-            open=-20,
-            extend=-10,
-            penalize_extend_when_opening=True,
-            one_alignment_only=True,
-        )
[…]
