### 2025-11-11 096fda1
Working on coverage and making sure single ended reads still work
