### 2025-11-11 812de57
Coverage needed a divide by 2 in paired-ended mode
