### 2025-11-11 c38e6d3
More updates to README.md with bioconda.
