### 2025-11-11 cd8aa76
Readability updates to README.md with bioconda.
