### 2025-11-12 c4290ad
Updating model, read names, and a few other details. Overall goal is to implement avg_seq_error in such a way that actually affects the output.
