### 2025-11-12 308cbd9
Expanding error model somewhat
