+++ neat/read_simulator/runner.py
@@ -80,22 +80,28 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
     reference_keys_with_lens = {key: len(value) for key, value in reference_index.items()}
 
     # We need sequencing errors to get the quality score attributes, even for the vcf
-    if options.error_model:
-        error_models = pickle.load(gzip.open(options.error_model))
-        error_model = error_models["error_model1"]
-    else:
-        # Use all the default values
-        error_model = SequencingErrorModel()
-
-    # Update error to user specified input
     if options.avg_seq_error:
-        error_model.average_error = options.avg_seq_error
+        average_error = options.avg_seq_error
+    elif options.error_model:
+        error_models = pickle.load(gzip.open(options.error_model))
+        average_error = error_models["error_model1"].average_error
+        # We just need the error value
+        del error_models
+    else:
+        # Use the default value
+        average_error = 0.009228843915252066
 
     # _LOG.debug('Sequencing error and quality score models loaded')
     # We need to estimate how many total errors to add
     total_reference_length = sum(reference_keys_with_lens.values())
-    total_errors = ceil(error_model.average_error * total_reference_length)
+    # This is an estimate due to some random effects
+    number_of_bases_in_analysis = total_reference_length * options.coverage
+    # Each base called has an "average_error" chance of being an error
+    total_errors = ceil(average_error * number_of_bases_in_analysis)
+    # Normalization gives the percent value for each item with a sum of all values being 1.0
     normalized_counts = {k: v / total_reference_length for (k, v) in reference_keys_with_lens.items()}
+    # Multiply the normalized count by the total errors. This gives the number of errors that should be
+    # introduced into the contig
     errors_per_contig = {k: ceil(v * total_errors) for (k, v) in normalized_counts.items()}
 
     count = 0
@@ -165,6 +171,14 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
     for contig in splits_files_dict:
         contig_index = contig_dict[contig]
         for ((start, length), splits_file) in splits_files_dict[contig].items():
+            block_percentage = length / reference_keys_with_lens[contig]
+            block_errors = errors_per_contig[contig] * block_percentage
+            estimated_number_of_reads = (length // options.read_len) * options.coverage
+            errors_per_read = round(block_errors / estimated_number_of_reads)
+            if errors_per_read < 1.0 and block_errors > 0:
+                # We know we need a few errors, but it's a small number total
+                if options.rng.random() < average_error:
+                    errors_per_read += 1
             current_output_dir = options.temp_dir_path / splits_file.stem
             current_output_dir.mkdir(parents=True, exist_ok=True)
             # Create local filenames based on fasta indexing scheme.
@@ -199,7 +213,7 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
                     target_regions_dict[contig],
                     discard_regions_dict[contig],
                     mutation_rate_dict[contig],
-                    errors_per_contig[contig],
+                    errors_per_read,
                 )
                 _LOG.info(f"Completed simulating contig {contig}.")
                 # TODO Remove if not needed
@@ -221,7 +235,8 @@ def read_simulator_runner(config: str, output_dir: str, file_prefix: str):
                     thread_input_variants,
                     thread_target_regions,
                     thread_discard_regions,
-                    thread_mutation_regions
+                    thread_mutation_regions,
+                    errors_per_read,
                 ))
             thread_idx += 1
 
