### 2025-11-12 85a9c51
Updated script to include paired-ended runs
