+++ BacterialWrapperScript/BacterialWrapperScript.py
@@ -3,12 +3,13 @@ import gzip
 import shutil
 import yaml
 import pysam
+import unittest
+import os
 
 from pathlib import Path
 from typing import List
-from neat.read_simulator.utils import Options, OutputFileWriter
 from Bio import bgzf
-from Bio.bgzf import BgzfWriter
+from Bio.bgzf import BgzfWriter, BgzfReader
 
 
 # Rearranges the bacterial chromosome by wrapping it around
@@ -42,31 +43,46 @@ def write_fasta_file(new_seq, bacteria_name, fasta_header):
     return fasta_file_name
 
 
+# Splits the coverage in half for the reference config file
 # Writes a yml configuration file for the newly rearranged chromosome's fasta sequence
 # This uses the default parameters for NEAT
 
 def write_config_file(ref_config_file, rearranged_seq_file, bacteria_name):
-    new_config_file_name = f"new_{bacteria_name}_config_test.yml"
+    new_config_file_name = f"test_new_{bacteria_name}_config_test.yml"
+    old_config_file_name = f"test_{bacteria_name}_config_test.yml"
 
-    with open(ref_config_file, 'r') as ref_file, open(new_config_file_name, 'w') as new_file:
+    with open(ref_config_file, 'r') as ref_file, open(new_config_file_name, 'w') as new_file, open(old_config_file_name, 'w') as old_file:
         
         for line in ref_file:
-            if line.find("reference:") == -1:
-                new_file.write(line)
-            else:
+            if line.find("reference:") != -1:
                 new_file.write(f"reference: {rearranged_seq_file}")
+                old_file.write(line)
+            elif line.find("coverage:") != -1:
+                if line.strip() == "coverage: .":
+                    new_coverage = 5.0
+                else:
+                    new_coverage = float((line.split(" "))[1].strip()) / 2
+        
+                new_file.write(f"coverage: {new_coverage}")
+                old_file.write(f"coverage: {new_coverage}")
+            else:
+                new_file.write(line)
+                old_file.write(line)
 
 
     ref_file.close()
     new_file.close()
+    old_file.close()
+
+    ref_file = old_file
 
     return new_config_file_name
 
 
 # Runs the NEAT read simulator using the given config file
 
-def run_neat(config_file):
-    subprocess.run(["neat", "read-simulator", "-c", config_file, "-o", config_file])
+def run_neat(config_file, output_dir, prefix):
+    subprocess.run(["neat", "read-simulator", "-c", config_file, "-o", output_dir + "/" + prefix])
 
 
 # General function for bacterial wrapper that calls all of the functions defined above
@@ -80,20 +96,20 @@ def bacterial_wrapper(reference_file, bacteria_name, ref_config_file):
 
     for line in f:
         if line[0] != ">":
-            orig_seq += line
+            orig_seq += line.strip()
         elif line.find("plasmid") != -1: # exclude plasmids from the sequence to be rearranged
             break
         
     f.close()
 
-    full_orig_seq = orig_seq.replace("\n", "")
[…]
