### 2025-11-13 e308fd0
Updating some parameters in error models
