### 2025-11-19 0edd6bc
Tables, descriptions, and more.
