### 2025-11-19 950ee30
A note about ploidy.
