### 2025-11-21 036a58e
fixing stitch and no fastq
