### 2025-11-29 069272c
Readability updates to README.
