### 2025-12-05 9d43fa3
Full Markov integration set of scripts.
