+++ neat/quality_score_modeling/__init__.py
@@ -1,0 +1,7 @@
+"""
+Load quality score modeling utilities for NEAT
+"""
+
+__all__ = ["build_markov_model"]
+
+from .markov_utils import build_markov_model
