+++ neat/model_quality_score/runner.py
@@ -1,0 +1,236 @@
+"""
+Runner for creating quality score models.
+
+This module implements the core logic for generating quality score models
+from input FASTQ files.
+
+The resulting models are saved as a gzip-compressed pickle file.
+"""
+
+import gzip
+import pickle
+import logging
+from pathlib import Path
+from typing import Iterable, List, Optional, Tuple
+
+import numpy as np
+
+from ..common import validate_input_path, validate_output_path
+from ..models import SequencingErrorModel, TraditionalQualityModel
+from ..model_sequencing_error.utils import parse_file
+from ..models.markov_quality_model import MarkovQualityModel
+from ..quality_score_modeling.markov_utils import build_markov_model
+
+__all__ = ["model_qual_score_runner"]
+
+_LOG = logging.getLogger(__name__)
+
+
+def _prepare_quality_scores_argument(
+    qual_scores: Iterable[int | float],
+) -> Tuple[List[int], bool]:
+    """Normalize the quality scores argument to a list of ints."""
+
+    if isinstance(qual_scores, int):
+        max_q = int(qual_scores)
+        return list(range(0, max_q + 1)), False
+
+    sorted_scores = sorted(int(x) for x in qual_scores)
+    max_q = max(sorted_scores)
+
+    return list(range(0, max_q + 1)), True
+
+
+def model_qual_score_runner(
+    files: List[str],
+    offset: int,
+    qual_scores: Iterable[int | float],
+    max_reads: int,
+    use_markov: bool,
+    overwrite: bool,
+    output_dir: str,
+    output_prefix: str,
+) -> None:
+    """Create and save a quality score model from FASTQ data.
+
+    Parameters
+    ----------
+    files:
+        A list of FASTQ file names.
+    offset:
+        Quality score offset (e.g. 33 for Sanger encoding).
+    qual_scores:
+        Either a single integer specifying the maximum quality score or an
+        iterable of integers specifying the allowed quality scores (binned
+        model).
+    max_reads:
+        Maximum number of reads to process from each input file. A value of
+        ``-1`` or ``float('inf')`` processes all reads.
+    use_markov:
+        If ``True``, construct a Markov chain–based quality model. If
+        ``False``, only the traditional quality model is produced.
+    overwrite:
+        If ``True``, allow an existing output file to be overwritten.
+    output_dir:
+        Directory to write the output model file.
+    output_prefix:
+        Prefix to use for the output file name. The file will be named
+        ``{output_prefix}.p.gz`` in the ``output_dir``.
+    """
+
+    if len(files) > 2:
+        _LOG.info("Only processing the first two input files")
+        files = files[:2]
+
+    # Validate input paths
+    for file in files:
+        validate_input_path(file)
+
+    _LOG.debug("Input files: %s", ", ".join(str(x) for x in files))
+    _LOG.debug("Quality offset: %d", offset)
+
+    final_quality_scores, binned = _prepare_quality_scores_argument(qual_scores)
+    _LOG.debug("Quality scores: %s", final_quality_scores)
+
+    # Determine maximum number of reads to process
+    if max_reads in (-1, None):
+        num_records_to_process = float("inf")
+
+    else:
+        num_records_to_process = max_reads
+    _LOG.debug(
+        "Maximum number of records to process: %s",
+        "all" if num_records_to_process == float("inf") else num_records_to_process,
+    )
+
+    # Validate output directory and file
+    validate_output_path(output_dir, is_file=False)
+    output_path = Path(output_dir)
+    output_file = output_path / f"{output_prefix}.p.gz"
+    validate_output_path(output_file, overwrite=overwrite)
+    _LOG.info("Writing output to: %s", output_file)
+
[…]
