### 2025-12-18 7c92f0f
Implemented CLI for wrapper
