### 2025-12-18 acef491
Implemented CLI for wrapper
