### 2026-01-24 9897d81
Made the naming for parallel_mode and mode consistent
