+++ neat/cli/cli.py
@@ -1,7 +1,7 @@
 """Implements command line interface used by the package."""
 
 __all__ = ['Cli', 'main', 'run']
-__version__ = "4.3.5"
+__version__ = "4.3.6"
 __author__ = "Joshua Allen"
 __email__ = "jallen17@illinois.edu"
 
