### 2026-01-30 7d6808f
Renaming to parallel_block_size and parallel_mode; fixing tests.
