+++ config_template/template_neat_config.yml
@@ -1,25 +1,27 @@
-## Template for gen_reads parallel
+## Template for NEAT's read-simulator (as of version 4.3.5, parallelization-friendly)
 ## Any parameter that is not required but has a default value will use the
 ## default value even if the variable is not included in the config. For
-## required items, they must be included in the config and the must be given a value.
+## required items, they must be included in the config and they must be given a value.
 ## All other items can be present or not. If present and the value is set to a single
 ## period, the variable will be treated as though it had been omitted. Please do
-## not modify this template, but instead make a copy in your working directory. Done this
-## way, you can run without even needing to declare -c.
+## not modify this template, but instead make a copy in your working directory.
+## Run with: neat read-simulator -c <config.yml> -o <out_dir> [-p <prefix>]
 
-# Absolute path to input reference fasta file
+# Absolute path to input reference FASTA file
 # type = string | required: yes
 reference: REQUIRED
 
-# Read length of the reads in the fastq output. Only required if @produce_fastq is set to true
-# type = int | required: no | default = 101
+# Read length of the reads in the FASTQ output. Only used if produce_fastq = true
+# type = int | required: no | default = 151
 read_len: .
 
-# Average Coverage for the entire genome.
-# type = float | required: no | default = 10.0
+# Average coverage for the entire genome
+# type = int | required: no | default = 10
 coverage: .
 
-# Absolute path to file with sequencing error model
+# Absolute path to file with sequencing error model or quality-score model
+# Error models are typically produced by neat model-seq-err (from FASTQ/BAM-like inputs)
+# Quality-score models can be produced by neat model-qual-score (optionally fit with --markov)
 # type = string | required: no | default: <NEAT_DIR>/neat/models/defaults/default_error_model.pickle.gz
 error_model: .
 
@@ -27,13 +29,11 @@ error_model: .
 # type = float | required = no | must be between 0.0 and 0.3
 avg_seq_error: .
 
-# This scales the quality scores to match the desired average sequencing error rate
-# specified by avg_seq_error.
+# Scale quality scores to match avg_seq_error
 # type: boolean | required = no | default = false
 rescale_qualities: .
 
-# This is the factor to add to the quality scores to get the ascii text version of the
-# score. The default follows the sanger quality offset
+# PHRED quality offset (e.g., 33 for Sanger)
 # type: int | required = no | default = 33
 quality_offset: .
 
@@ -41,126 +41,99 @@ quality_offset: .
 # type = int | required = no | default = 2
 ploidy: .
 
-# Absolute path to vcf file containing variants that will always be included, regardless
-# of genotype and filter. You can pre-filter your vcf for these fields before inputting it
-# if this is not the desired behavior.
+# Absolute path to VCF file containing variants that will always be included
 # type: string | required = no
 include_vcf: .
 
-# Absolute path to bed file containing reference regions that the simulation
-# should target.
+# Absolute path to BED file containing reference regions that the simulation should target
 # type = string | required = no
 target_bed: .
 
-# Scalar value for coverage in regions outside the targeted bed. Example 0.5
-# would get you roughly half the coverage as the on target areas. Default is
-# 0 coverage in off-target regions. Number should be a float in decimal.
-# type: float | required = no | default = 0.00
-off_target_scalar: .
-
-# Absolute path to bed file containing reference regions that the simulation
-# should discard.
+# Absolute path to BED file containing reference regions that the simulation should discard
 # type = string | required = no
 discard_bed: .
 
-# Absolute path to the mutation model pickle file. Omitting this value will cause
-# NEAT to use the default model, with some standard parameters, and generally uniform biases.
[…]
