### 2026-02-01 8459c8b
Fixed bugs with config files and output stitching
