### 2026-02-01 a73ce4c
Fixed bugs with config files and output stitching
