+++ README.md
@@ -81,26 +81,24 @@ $ git clone git@github.com:ncsa/NEAT.git
 $ cd NEAT
 ```
 
-A quick form of installation uses `bioconda`. Once `conda` is installed, the following command can be run for easy setup. 
-In the NEAT repo, at the base level is the environment.yml file you will need. Change directories into the neat repository
+A quick form of installation uses `bioconda`. You must run these commands inside the NEAT project directory.
+
+```bash
+(base) $ conda create -n neat -c conda-forge -c bioconda neat
+(base) $ conda activate neat
+(neat) $ neat --help # tests that NEAT has installed correctly
+```
+
+Alternatively, instead of the `bioconda` method, you can use the `poetry` module in build a wheel file, which can then be `pip` installed.
+
+Once `conda` is installed, the following command can be run for easy setup. 
+In the NEAT repository, at the base level is the `environment.yml` file you will need. Change directories into the NEAT repository
 and run:
 
 ```bash
 (base) $ conda env create -f environment.yml
 (base) $ conda activate neat
 (neat) $ poetry install
-(neat) $ neat --help # tests that NEAT has installed correctly
-```
-
-Alternatively, instead of the `bioconda` method, you can use the `poetry` module in build a wheel file, which can then be `pip` installed.
-
-You will need to run these commands from within the NEAT directory:
-
-```bash
-(base) $ conda env create -f environment.yml
-(base) $ conda activate neat
-(neat) $ poetry build
-(neat) $ pip install dist/neat*whl
 (neat) $ neat --help # tests that NEAT has installed correctly
 ```
 
