### 2026-02-22 123429e
Testing CLI and new tests for error and mutation models.
