+++ tests/test_models/test_error_and_mut_models.py
@@ -43,8 +43,7 @@ def test_mutation_model_generate_snv_trinuc():
 
 def test_sequencing_error_model_zero_error_returns_none_or_empty():
     """
-    avg_seq_error == 0 should yield no errors. Some versions return just the list,
-    others return a (list, padding) tuple — accept both.
+    avg_seq_error == 0 should yield no errors.
     """
     rng = default_rng(4)
     sem = SequencingErrorModel(avg_seq_error=0.0)
@@ -89,3 +88,105 @@ def test_sequencing_error_model_basic_snvs_only():
         assert hasattr(e, "location")
         assert hasattr(e, "ref")
         assert hasattr(e, "alt")
+
+def test_mutation_model_insertion_reproducible_with_seed():
+    """Same seed and inputs should give the same insertion (length and alt)."""
+    rng1 = default_rng(123)
+    rng2 = default_rng(123)
+    m = MutationModel()
+    ref = Seq("ACGT")
+
+    ins1 = m.generate_insertion(location=10, ref=ref, rng=rng1)
+    ins2 = m.generate_insertion(location=10, ref=ref, rng=rng2)
+
+    assert isinstance(ins1, Insertion)
+    assert isinstance(ins2, Insertion)
+    assert ins1.length == ins2.length
+    assert str(ins1.alt) == str(ins2.alt)
+
+
+def test_mutation_model_deletion_reproducible_with_seed():
+    """Same seed and inputs should give the same deletion object shape."""
+    rng1 = default_rng(456)
+    rng2 = default_rng(456)
+    m = MutationModel()
+
+    del1 = m.generate_deletion(location=25, rng=rng1)
+    del2 = m.generate_deletion(location=25, rng=rng2)
+
+    assert isinstance(del1, Deletion)
+    assert isinstance(del2, Deletion)
+    assert del1.length == del2.length
+    assert del1.position1 == del2.position1
+
+
+def test_mutation_model_snv_does_not_keep_reference_base():
+    """
+    For a given trinucleotide, the generated SNV should change the central base.
+    """
+    rng = default_rng(7)
+    m = MutationModel()
+    trinuc = Seq("ACA")
+    central = str(trinuc[1])
+
+    snv = m.generate_snv(trinucleotide=trinuc, reference_location=100, rng=rng)
+    assert isinstance(snv, SingleNucleotideVariant)
+    assert snv.alt in ["A", "C", "G", "T"]
+    assert snv.alt != central
+
+
+def test_traditional_quality_model_reproducible_with_seed():
+    """Quality model should be deterministic given the same RNG state."""
+    rng1 = default_rng(8)
+    rng2 = default_rng(8)
+    qm = TraditionalQualityModel(average_error=0.01)
+
+    qs1 = qm.get_quality_scores(model_read_length=151, length=100, rng=rng1)
+    qs2 = qm.get_quality_scores(model_read_length=151, length=100, rng=rng2)
+
+    assert np.array_equal(qs1, qs2)
+
+
+def test_sequencing_error_model_reproducible_with_seed():
+    """Error placement should be deterministic given the same RNG state."""
+    sem = SequencingErrorModel(avg_seq_error=0.05)
+    ref = SeqRecord(Seq("ACGT" * 30), id="chr1")
+    quals = np.array([30] * len(ref), dtype=int)
+
+    rng1 = default_rng(9)
+    rng2 = default_rng(9)
+
+    introduced1, pad1 = sem.get_sequencing_errors(
+        padding=20, reference_segment=ref, quality_scores=quals, rng=rng1
+    )
+    introduced2, pad2 = sem.get_sequencing_errors(
+        padding=20, reference_segment=ref, quality_scores=quals, rng=rng2
+    )
+
[…]
