### 2026-02-22 f3ade95
An RNG-related bug covered by tests was solved with this change.
