### 2026-02-23 5a4d347
Markov quality score model changes copied in here.
