+++ neat/models/quality_score_model.py
@@ -1,426 +1,0 @@
-import pysam
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-from matplotlib.patches import Patch
-import seaborn as sns
-from scipy.stats import ttest_ind, ttest_rel, f_oneway, norm, levene, shapiro
-from sklearn.utils import resample
-import pathlib
-import pickle
-
-
-def make_qual_score_list(bam_file):
-    """Takes an input BAM file and creates lists of quality scores. This becomes a data frame, which will be
-    pre-processed for Markov chain analysis."""
-
-    index = f"{bam_file}.bai"
-
-    if not pathlib.Path(index).exists():
-        print("No index found, creating one.")
-
-        pysam.index(bam_file)
-
-    file_to_parse = pysam.AlignmentFile(bam_file, "rb", check_sq=False)
-    num_recs = file_to_parse.count()
-    print(f"{num_recs} records to parse")
-
-    modulo = round(num_recs / 9)
-
-    qual_list = []
-    i = 0
-    j = 0
-
-    def print_update(number, factor, percent):
-
-        if number % factor == 0:
-            percent += 10
-            print(f"{percent}% complete", end="\r")
-
-        return percent
-
-    print("Parsing file")
-
-    for item in file_to_parse.fetch():
-
-        if item.is_unmapped or "S" in item.cigarstring:
-            i += 1
-            j = print_update(i, modulo, j)
-
-            continue
-
-        # mapping quality scores
-
-        align_qual = item.query_alignment_qualities
-
-        # append to master lists
-
-        qual_list.append(align_qual)
-        i += 1
-        j = print_update(i, modulo, j)
-
-    print(f"100% complete")
-    file_to_parse.close()
-
-    quality_df = pd.DataFrame(qual_list)  # turn list of lists into a dataframe
-    quality_df = quality_df.fillna(0)  # pre-processing - fill in missing data
-
-    # filters to process outliers
-
-    quality_df[quality_df > 40] = 40
-    quality_df[quality_df < 0] = 0
-
-    return quality_df
-
-
-def estimate_transition_probabilities(std_dev):
-    """Takes a standard deviation as an input and generates the transition probabilities with a normal
-    distribution that can be used to represent a Markov process."""
-
-    # define the probabilities for transition states based on a normal distribution
-
-    transition_probs = {
-        -3: norm.pdf(-3, 0, std_dev),
-        -2: norm.pdf(-2, 0, std_dev),
-        -1: norm.pdf(-1, 0, std_dev),
-        0: norm.pdf(0, 0, std_dev),
-        1: norm.pdf(1, 0, std_dev),
-        2: norm.pdf(2, 0, std_dev),
-        3: norm.pdf(3, 0, std_dev),
-    }
-
-    # normalize the probabilities to sum to 1
-
-    total_prob = sum(transition_probs.values())
-
-    for k in transition_probs:
-        transition_probs[k] /= total_prob
-
-    return transition_probs
-
-
-def apply_markov_chain(quality_df, noise_level=10, std_dev=2):
-    """Takes a data frame representing quality scores by position along a read and parameters to increase
-    variability in the Markov process as inputs and generates predictions based on an ergodic Markov chain.
-    Generates a data frame with simulated reads."""
-
-    transition_probs = estimate_transition_probabilities(std_dev)
-    num_rows, num_cols = quality_df.shape
-
-    count = 0
-    markov_preds = []
-
-    for row in quality_df.iterrows():
-
-        qualities = row[1].values
[…]
