### 2026-02-26 d3542ac
Adjusting coverage calculations
