+++ neat/read_simulator/utils/stitch_outputs.py
@@ -17,7 +17,7 @@ from neat.read_simulator.utils import OutputFileWriter
 
 _LOG = logging.getLogger(__name__)
 
-def concat(files_to_join: List[Path], dest_file: BgzfWriter) -> None:
+def concat(files_to_join: List[Path], dest_file: gzip.GzipFile) -> None:
     if not files_to_join:
         # Nothing to do, and no error to throw
         _LOG.warning(f"Concat called but there are no files to join: {files_to_join}" )
