### 2026-02-27 b0ad46a
README updates including Markov stuff.
