### 2026-03-05 e94e785
Transition matrices better accounted for.
