+++ neat/models/markov_quality_model.py
@@ -1,21 +1,16 @@
 """
-Empirical position-specific quality score model for NEAT.
+Position-specific quality score model for NEAT.
 
 The model consists of an initial distribution giving the probability of observing each
-quality score at the first position of a read and a sequence of per-position empirical
-distributions giving the probability of observing each quality score at each subsequent
-position.
+quality score at the first position of a read, per-position marginals, and per-position
+transition matrices.
 
-At generation time, the model samples a quality score independently at
-each position from the corresponding empirical distribution.
-
-Although this model exists outside of ``neat.models.error_models``, it
-parallels the ``TraditionalQualityModel`` class from NEAT’s core and can
-be used interchangeably when constructing quality models.
+If transition_distributions is None, the model independently samples from
+per-position marginals.
 """
 
 from dataclasses import dataclass
-from typing import Dict, List
+from typing import Dict, List, Optional
 
 import numpy as np
 
@@ -23,8 +18,7 @@ __all__ = ["MarkovQualityModel"]
 
 @dataclass
 class MarkovQualityModel:
-    """Empirical, position-specific quality score model.
-
+    """
     Parameters
     ----------
     initial_distribution:
@@ -40,12 +34,15 @@ class MarkovQualityModel:
     read_length:
         The length of reads used during training determines how many
         position-specific distributions exist.
+    transition_distributions:
+        Optional list trans[i][q_prev][q_next] = count/prob.
     """
 
     initial_distribution: Dict[int, float]
     position_distributions: List[Dict[int, float]]
     max_quality: int
     read_length: int
+    transition_distributions: Optional[List[Dict[int, Dict[int, float]]]] = None
 
     def __post_init__(self) -> None:
 
@@ -58,24 +55,20 @@ class MarkovQualityModel:
             )
 
         self.initial_distribution = {
-            int(k): float(v) / total_init
-            for k, v in self.initial_distribution.items()
+            int(k): float(v) / total_init for k, v in self.initial_distribution.items()
         }
 
-        # Normalize each position-specific distribution
+        # Normalize per-position marginals
         norm_positions: List[Dict[int, float]] = []
 
         for dist in self.position_distributions:
             total = float(sum(dist.values()))
 
             if total <= 0:
-                # Fall back to a single mass at quality 0
                 norm_positions.append({0: 1.0})
                 continue
 
-            norm_positions.append(
-                {int(k): float(v) / total for k, v in dist.items()}
-            )
+            norm_positions.append({int(k): float(v) / total for k, v in dist.items()})
 
         self.position_distributions = norm_positions
 
@@ -97,6 +90,38 @@ class MarkovQualityModel:
             vals = [dist[k] for k in keys]
             self._values_by_pos.append(np.asarray(keys, dtype=int))
             self._probs_by_pos.append(np.asarray(vals, dtype=float))
+
+        # Normalize and cache transitions
+        self._has_transitions = False
+        self._trans_values_by_pos: List[Dict[int, np.ndarray]] = []
+        self._trans_probs_by_pos: List[Dict[int, np.ndarray]] = []
+
+        if self.transition_distributions is not None:
+            if len(self.transition_distributions) not in (0, max(0, self.read_length - 1)):
+                raise ValueError(
+                    "transition_distributions must have length read_length-1 "
+                    f"(expected {max(0, self.read_length - 1)}, got {len(self.transition_distributions)})."
+                )
+
+            # Normalize each row q_prev -> distribution over q_next
+            self._has_transitions = len(self.transition_distributions) > 0
+
+            for pos_trans in self.transition_distributions:
+                cache_vals: Dict[int, np.ndarray] = {}
[…]
