+++ neat/quality_score_modeling/markov_utils.py
@@ -1,17 +1,18 @@
 """
-Utility functions for building an empirical quality score model.
+Utility functions for building a position-specific Markov quality score model.
 
 The functions in this module extract per-read quality scores from FASTQ files
-and compute the statistics required by the ``MarkovQualityModel``.
+and compute the information required by the ``MarkovQualityModel``.
 """
 
 import logging
+from bisect import bisect_right
 from collections import defaultdict
 from pathlib import Path
-from typing import Dict, Iterable, List, Tuple
+from typing import Dict, Iterable, List, Optional, Tuple
 
-from ..model_sequencing_error.utils import convert_quality_string
 from ..common import open_input
+from ..model_sequencing_error.utils import convert_quality_string
 
 _LOG = logging.getLogger(__name__)
 
@@ -19,37 +20,47 @@ __all__ = [
     "read_quality_lists",
     "compute_initial_distribution",
     "compute_position_distributions",
+    "compute_transition_distributions",
     "build_markov_model",
 ]
+
+
+def _down_bin_quality(q: int, allowed: List[int]) -> int:
+    """
+    Map q to the greatest allowed value <= q (down-binning).
+    If q is below the smallest allowed, map to allowed[0].
+    """
+
+    if not allowed:
+        return int(q)
+
+    q = int(q)
+    i = bisect_right(allowed, q) - 1
+
+    if i < 0:
+        return int(allowed[0])
+
+    return int(allowed[i])
 
 
 def read_quality_lists(
     files: Iterable[str],
     max_reads: int,
     offset: int,
+    allowed_quality_scores: Optional[Iterable[int]] = None,
 ) -> Tuple[List[List[int]], int]:
     """Read per-read quality scores from one or more FASTQ files.
 
-    Parameters
-    ----------
-    files:
-        An iterable of FASTQ filenames. Only the first ``max_reads`` reads
-        will be consumed from each file. If a file contains fewer than
-        ``max_reads`` reads, the entire file is processed.
-    max_reads:
-        Maximum number of reads per file to consume. A value of ``-1``
-        or ``float('inf')`` indicates that all reads in the file should be
-        processed.
-    offset:
-        Numeric quality score offset (usually 33 for Sanger encoding).
-
-    Returns
-    -------
-    Tuple[List[List[int]], int]
-        A tuple containing the list of quality score lists and the read
-        length inferred from the first record. If no reads are found this
-        function returns an empty list and zero.
+    Returns (qualities, read_length). If no reads are found, returns ([], 0).
     """
+
+    allowed_sorted: Optional[List[int]] = None
+
+    if allowed_quality_scores is not None:
+        allowed_sorted = sorted({int(x) for x in allowed_quality_scores})
+
+        if not allowed_sorted:
+            allowed_sorted = None
 
     qualities: List[List[int]] = []
     read_length = 0
@@ -71,32 +82,35 @@ def read_quality_lists(
 
             while reads_read < reads_to_parse:
 
-                # FASTQ format: four lines per record (header, sequence, plus, quality)
+                # FASTQ format: four lines per record
                 header = fq_in.readline()
 
                 if not header:
                     break  # end of file
 
-                seq = fq_in.readline()
-                plus = fq_in.readline()
+                _ = fq_in.readline()  # seq
+                _ = fq_in.readline()  # plus
                 qual = fq_in.readline()
 
                 if not qual:
-                    break  # truncated file
+                    break
 
-                quality_scores = convert_quality_string(qual.strip(), offset)
+                qlist = convert_quality_string(qual.strip(), offset)
 
                 if not read_length:
[…]

[…]
