### 2026-03-05 ff13ea9
Formatting in template_neat_config.yml
