### 2026-03-05 0b69f8d
Delete version.py (we already have a 4.3.6 and version tracker now)
