+++ paper.md
@@ -1,0 +1,122 @@
+---
+title: 'Enhancing short-read sequencing simulation: Updates to NEAT'
+
+tags:
+  - Python
+  - genomics
+  - DNA sequencing
+  - simulation
+
+authors:
+  - name: Joshua M. Allen
+    equal-contrib: true
+    affiliation: 1
+  - name: Keshav R. Gandhi
+    orcid: 0009-0000-1718-1862
+    equal-contrib: true
+    corresponding: true
+    affiliation: "1, 2"
+  - name: Raghid Alhamzy
+    affiliation: 1
+  - name: Yash Wasnik
+    affiliation: 3    
+  - name: Christina E. Fliege
+    affiliation: 1
+    corresponding: true
+
+affiliations:
+ - name: National Center for Supercomputing Applications, Genomics Group, Urbana, IL, USA, 61801, 
+   index: 1
+   
+ - name: University of Illinois at Chicago, Chicago, IL, USA, 60607
+   index: 2
+
+ - name: Blue Health Intelligence, Chicago, United States, 60611
+   index: 3
+   
+date: 5 March 2026
+
+bibliography: paper.bib
+---
+
+# Summary
+
+While the field of genomics has advanced significantly with the advent of high-throughput sequencing technologies, challenges related to the availability, complexity, and variability of this data can introduce difficulty to the development and validation of computational tools. Simulated short-read sequencing datasets provide researchers a method to obtain reproducible, verified data to test algorithms and benchmark software. Simulations also avoid the limitations of working with real data, including the cost of genomic sequencing, time to process sequencing data, accessibility of data, and protection of privacy for human datasets. Ideally, these datasets mimic the properties of real sequencing datasets. From introducing specific patterns of sequencing errors to modeling localized regions of mutations, realistic datasets are necessary to evaluate the accuracy and robustness of downstream alignment, variant-calling, and analysis pipelines. Sequencing parameters such as guanine-cytosine (GC) content, repeat structure, and genome complexity vary across species, and it is important for simulated reads to reflect these species-specific characteristics. The NExt-generation sequencing Analysis Toolkit (NEAT) is an open-source Python package that creates such simulated sequencing datasets and was originally released as NEAT 2.0 in 2016. Here, we describe ongoing updates released collectively as NEAT 4.3, including new features and increased speed, accuracy, and usability.
+
+# Statement of need
+
+Developing and validating methods for read alignment, variant calling, structural variant detection, and other analyses depends on access to genomic data with known ground truth. As a result, many research groups rely on simulated reads, and it can be useful to vary the parameters of the sequencing process itself in order to demonstrate how analysis pipelines may vary in performance across different coverages, read lengths, ploidies, and more. NEAT addresses this need as an open-source Python package that can easily integrate within existing bioinformatics workflows—its simulations account for a wide range of sequencing parameters (e.g., coverage, fragment length, sequencing error models, mutational frequencies, ploidy, etc.) and allow users to customize their sequencing data [@Stephens:2016].
+
+NEAT is designed to simulate short reads. NEAT is adaptable to different sequencing machines with user-specific, custom error models and the capability to handle single-base substitutions, insertions, deletions, and larger structural variants. Unlike simulators that rely solely on fixed error profiles, NEAT can learn empirical mutation and sequencing models from real datasets and use these models to generate realistic sequencing data, providing outputs in several common bioinformatics file formats, including FASTQ, Binary Alignment Map (BAM), and Variant Call Format (VCF) files. NEAT provides utilities for producing BAM and VCF files, facilitating benchmarking of full bioinformatics pipelines.
+
[…]
