+++ paper.bib
@@ -1,0 +1,280 @@
+@article{Stephens:2016,
+    author = {Stephens, Z. D. and Hudson, M. E. and Mainzer, L. S. and Taschuk, M. and Weber, M. R. and Iyer, R. K.},
+    title = {Simulating Next-Generation Sequencing Datasets from Empirical Mutation and Sequencing Models},
+    journal = {PLOS ONE},
+    year = {2016},
+    volume = {11},
+    number = {11},
+    pages = {e0167047},
+    doi = {10.1371/journal.pone.0167047}
+}
+
+@article{Benjamini:2012,
+    author = {Benjamini, Y. and Speed, T. P.},
+    title = {Summarizing and correcting the GC content bias in high-throughput sequencing},
+    journal = {Nucleic Acids Research},
+    year = {2012},
+    volume = {40},
+    number = {10},
+    pages = {e72},
+    doi = {10.1093/nar/gks001}
+}
+
+@article{Ross:2013,
+    author = {Ross, M. G. and Russ, C. and Costello, M. and Hollinger, A. and Lennon, N. J. and Hegarty, R. and Nusbaum, C. and Jaffe, D. B.},
+    title = {Characterizing and measuring bias in sequence data},
+    journal = {Genome Biology},
+    year = {2013},
+    volume = {14},
+    number = {5},
+    pages = {R51},
+    doi = {10.1186/gb-2013-14-5-r51}
+}
+
+@article{Escalona:2016,
+    author = {Escalona, M. and Rocha, S. and Posada, D.},
+    title = {A comparison of tools for the simulation of genomic next-generation sequencing data},
+    journal = {Nature Reviews Genetics},
+    year = {2016},
+    volume = {17},
+    number = {8},
+    pages = {459--469},
+    doi = {10.1038/nrg.2016.57}
+}
+
+@article{Zhao:2017,
+    author = {Zhao, M. and Liu, D. and Qu, H.},
+    title = {Systematic review of next-generation sequencing simulators: computational tools, features and perspectives},
+    journal = {Briefings in Functional Genomics},
+    year = {2017},
+    volume = {16},
+    number = {3},
+    pages = {121--128},
+    doi = {10.1093/bfgp/elw012}
+}
+
+@article{Alosaimi:2020,
+    author = {Alosaimi, Shatha and Bandiang, Armand and {van Biljon}, Noelle and Awany, Denis and Thami, Prisca K. and Tchamga, Milaine S. S. and Kiran, Anmol and Messaoud, Olfa and Hassan, Radia Ismaeel Mohammed and Mugo, Jacquiline and Ahmed, Azza and Bope, Christian D. and Allali, Imane and Mazandu, Gaston K. and Mulder, Nicola J. and Chimusa, Emile R.},
+    title = {A broad survey of DNA sequence data simulation tools},
+    journal = {Briefings in Functional Genomics},
+    year = {2020},
+    volume = {19},
+    number = {1},
+    pages = {49--59},
+    doi = {10.1093/bfgp/elz033}
+}
+
+@article{Milhaven:2023,
+    author = {Milhaven, M. and Pfeifer, S. P.},
+    title = {Performance evaluation of six popular short-read simulators},
+    journal = {Heredity},
+    year = {2023},
+    volume = {130},
+    pages = {55--63},
+    doi = {10.1038/s41437-022-00577-3}
+}
+
+@article{Schmeing:2021,
+    author = {Schmeing, S. and Robinson, M. D.},
+    title = {ReSeq simulates realistic Illumina high-throughput sequencing data},
[…]

[…]
