+++ .github/workflows/paper-pdf.yml
@@ -1,0 +1,33 @@
+name: Paper PDF
+on:
+  push:
+    paths:
+      - paper.md
+      - paper.bib
+      - .github/workflows/paper-pdf.yml
+
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Paper Draft
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Open Journals PDF Generator
+        uses: openjournals/openjournals-draft-action@v.1.0
+        with:
+          journal: joss
+          # This should be the path to the paper within your repo.
+          paper-path: paper.md
+      - name: Upload
+        uses: actions/upload-artifact@v4
+        with:
+          name: paper
+          # This is the output path where Pandoc will write the compiled
+          # PDF. Note, this should be the same directory as the input
+          # paper.md
+          path: paper.pdf
+
+      - name: Debug list files
+        run: ls -R
+      # Double-check to find the actual path of the generated PDF
