### 2026-03-05 aef2e25
Minor paper edits and README.md edits.
