+++ paper.md
@@ -13,6 +13,7 @@ authors:
     affiliation: 1
   - name: Keshav R. Gandhi
     orcid: 0009-0000-1718-1862
+    email: krg3@uic.edu
     equal-contrib: true
     corresponding: true
     affiliation: "1, 2"
@@ -22,13 +23,14 @@ authors:
     affiliation: 3    
   - name: Christina E. Fliege
     affiliation: 1
+    email: cfliege2@illinois.edu
     corresponding: true
 
 affiliations:
  - name: National Center for Supercomputing Applications, Genomics Group, Urbana, IL, USA, 61801, 
    index: 1
    
- - name: University of Illinois at Chicago, Chicago, IL, USA, 60607
+ - name: University of Illinois at Chicago, Chicago, IL, USA, 60607,
    index: 2
 
  - name: Blue Health Intelligence, Chicago, United States, 60611
@@ -51,8 +53,6 @@ NEAT is designed to simulate short reads. NEAT is adaptable to different sequenc
 
 After the original release of NEAT 2.0, most scripts have undergone significant ongoing changes since 2020. Upgrading to Python 3 brought NEAT 4.3 up to modern coding standards and allowed it to use standard Python libraries in order to streamline the code and improve its maintainability. The toolkit is optimized for both speed and accuracy, and new features have been implemented. A summary of algorithmic changes is provided in **Table 1**, and a summary of updates to NEAT’s software robustness and user experience is provided in **Table 2**.
 
-\newpage
-
 # Tables
 
 ### Table 1. Algorithmic Improvements and Methodological Changes
@@ -62,40 +62,63 @@ After the original release of NEAT 2.0, most scripts have undergone significant 
 +===+=======================+================================================================+======================================================+
 | 1 | BAM File Generation   | File generation was tightly integrated with all NEAT processes | BAM creation isolated from core functions            |
 +---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
+|   |                       |                                                                |                                                      |
++---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
 | 2 | GC Bias Computation   | Used a custom script for GC bias calculation                   | Bias can be accounted for with BED targeting         |
++---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
+|   |                       |                                                                |                                                      |
 +---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
 | 3 | Ploidy Simulation     | Limited to diploid organisms in practice                       | Supports polyploidy                                  |
 +---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
+|   |                       |                                                                |                                                      |
++---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
 | 4 | Read Generation       | Sliding-window approach to generate reads                      | A new form of coordinate-based read selection        |
++---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
+|   |                       |                                                                |                                                      |
 +---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
 | 5 | Read Quality Modeling | Markov-based model                                             | Revised binning method (optional Markov-based model) |
 +---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
+|   |                       |                                                                |                                                      |
++---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
 | 6 | Variant Handling      | Limited introduction of new variant types                      | A modular design with generic variant handling       |
++---+-----------------------+----------------------------------------------------------------+------------------------------------------------------+
+|   |                       |                                                                |                                                      |
[…]
