+++ .github/workflows/paper-pdf.yml
@@ -17,15 +17,14 @@ jobs:
         uses: openjournals/openjournals-draft-action@v.1.0
         with:
           journal: joss
-          # This should be the path to the paper within your repo.
+          # This is the path to the paper within the repo
           paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
+          # PDF (should be the same directory as the input paper.md)
           path: paper.pdf
 
       - name: Debug list files
