### 2026-03-05 #9056 @keshav-gandhi
Hello @csoneson, @bricoletc, @chenyenchung, and @erik-whiting!

We have officially gained approval to submit our revisions to NEAT. The current paper is available on the `paper` branch alongside the most up-to-date code and documentation.

Thank you so much for your careful reviews and patience as we have rolled out the software, fixed bugs, and continued to refine NEAT! We have taken time to make significant changes to the codebase and have been improving and testing NEAT after both the initial submission to JOSS and after receiving your feedback.

We have worked carefully to address all of your constructive feedback and appreciate all of the time and effort you spent helping us improve our software!

Below each quote are our responses to your feedback:

----------------------------------------------------------------------------------------------------

@bricoletc

> More explicit mentions to improve general clarity of the paper:
> L.22: 'Since the original release' please state version number (presumably 2.0?)
> L.29: common file formats, please state explicitly e.g. '(FASTQ, SAM/BAM, VCF)'

Response: Thank you! We have included these changes.

> Data types: what sequencing technologies/read lengths do you support, just Illumina short reads, or long reads, and if so with what error models? Please state this in your paper.

Response: These technologies are now mentioned explicitly.

Addition: “NEAT is designed to simulate short reads. NEAT is adaptable to different sequencing machines with user-specific, custom error models and the capability to handle single-base substitutions, insertions, deletions, and larger structural variants.”

> As mentioned by another reviewer, you're missing mention of state of the field. What other tools for read simulation exist, how does your software compare to them? This should be relatively easy as some high-profile papers seem to have used NEAT, and benchmark papers have shown NEAT performs well - it's a question of stating this and what else exists out there.

Response: There is now a unique “State of the field” section consisting of multiple paragraphs that highlight other tools for read simulation. Findings from several unique reviews and benchmarking studies of read simulation tools are also mentioned.

> Authorship question:
> I noticed there a lot of commits since, say, 2022, by user YashWasnik7; should this person be an author on your paper?

Response: We have added Raghid Alhamzy and Yash Wasnik as co-authors on the paper. Raghid’s contributions were focused on the BAM creation scripts, and Yash’s contributions were focused on setting up GitHub Actions for automated testing.

> Table 1:
> Can you define GC bias in the context of read simulation, and state why was it deprecated?

Response: We have now defined GC bias in the paper and clarified why it was deprecated. Furthermore, we found that in practice, the GC bias script had minimal impact on any final outputs in NEAT 2.0.

Addition: “GC bias computation (2) models the tendency of sequencing instruments to produce uneven coverage as a function of a local genomic region’s GC content, often underrepresenting regions of particularly high or low GC levels [@Benjamini:2012; @Ross:2013]. To streamline the codebase, NEAT 4.3 removes the GC-bias module. Instead, NEAT allows users to control coverage patterns via Browser Extensible Data (BED) region targeting. By default, this approach sets coverage outside targeted intervals to zero, enabling users to directly specify where reads are sampled.”

> 'Variant insertion (6) was updated to preserve genotype data in the final simulated Variant Call Format (VCF) file,' what does this mean, previously insertions were not recorded in output VCF?

Response: NEAT 2.0 only accounts for a VCF file’s `chrom`, `alt`, and `ref` positions. We now account for more information (e.g., allele frequencies, genotypes, etc.), added a genotype reader, and preserved more information.

Addition: “Finally, variant insertion (7) handles how known mutations are integrated into simulated genomes and reported in VCF files. NEAT offers an option for a user to input a VCF file of variants to be incorporated into the simulated reads. NEAT 4.3 now preserves data in the per-sample genotype and allele-frequency VCF fields in the final simulated VCF files.”

> Table 2:
> I don't know that tests are an 'improvement to user experience' as in the title of Table 2 of your paper. Maybe rename table to 'Improvements to software robustness and user experience'

Response: The table has been appropriately renamed to “Improvements to Software Robustness and User Experience.”

> Installation:
> The README, and Table 2, states you can install NEAT through poetry/pip, and preferably in a conda environment.
> I found a few things confusing:
>
> Seeing as you recommend conda/poetry/pip in your README, I immediately thought of distributing your software through bioconda, and instructing to install through that. In fact there is a bioconda release of NEAT, seems up to date too (v4.3):
> https://bioconda.github.io/recipes/neat/README.html#package-package%20&#x27;neat&#x27;
> --> Why not state this in README and the paper, and install though there?

Response: One of our users installed a bot to install NEAT via Bioconda. We have added more information to the README and manuscript about Bioconda as an alternate, quick method of installation.

> Why maintain both environment.yml and pyproject.toml? I get that you could maintain both in theory, but there are redundancies, e.g. installing pysam through both.

Response: We note that the use of Poetry requires `pyproject.toml` to be maintained. This is because Poetry (as it is used in NEAT) requires a virtual environment to function. We note that `environment.yml` has been greatly simplified and streamlined to remove any redundancies. This decision was also made in support of Ubuntu, which installs a system Python; Poetry requires a Python version, but Poetry does not install Python on its own. This leaves Conda as a convenient way to both account for this problem and serve as a virtual environment for Poetry.

> Testing:
> I like that you added tests in the new version of NEAT, this is very important.
>
> Seems to me like you have unit tests, through pytest; please include how to run those tests in [CONTRIBUTING.md](http://contributing.md/)
> pytest should probably a development dependency, e.g. in pyproject.toml
> Your github actions seems not to run the unit tests, but to run mock instances of read simulation? (https://github.com/ncsa/NEAT/blob/main/.github/workflows/python-app.yml)

Response: The GitHub is now configured to run the unit tests alongside the read simulation CLI integration test with pushes and pull requests. The test scripts are automated, and there is now documentation at `CONTRIBUTING.md` that describes how to run the tests. Similarly, we now add `pytest` as a dependency in `pyproject.toml`.

> Timing of publication:
>
> Your README states that NEAT v4.0 and above is not stable, as it's slow to run. So should your paper mention the stable version (v3.4), or in the paper, state that v4.2/v4.3 isn't quite stable. By the way when do you expect to have this stable?

Response: We have addressed this concern by rolling out a parallelized, stable version of NEAT 4.3 and mentioning this in `README.md`. There are also benchmarking tables in `README.md` that can be used for further reference.

----------------------------------------------------------------------------------------------------

@chenyenchung

> Overall
>
> I like the brevity and clarity of how the manuscript was written, and find the installation and usage instruction in the repository easy to follow. However, I found it not immediately obvious what the tool was designed for due to the rapid expansion of next-generation sequencing applications.
>
> Suggestion 1: Given the main use case appears to be simulation of short-read whole-genome or whole-exome sequencing, can this be explicit in the title, summary, and statement of need?

Response: The new title is “Enhancing Short-Read Sequencing Simulation: Updates to NEAT” and the summary and statement of need now specifically mention short reads.

> Functionality
>
> The instruction is fine and streamlined as it is, but I have similar confusion as @bricoletc:
>
> Question 1: Why are both poetry and conda necessary (are they)?

Response: We describe this above when referencing feedback from @brioletc.

> Question 2: Is the bioconda package working properly? If so, it could be a more user-friendly way of distribution.

Response: We have updated `README.md` and manuscript to reflect the Bioconda installation.

> And, given the current environment.yml, due to the specification of libgcc version, the installation can only be done per instruction on Linux but not MacOS (I don't have access to a Windows machine, so I don't know if it works on Windows and WSL).
>
> Suggestion 2: Please specify the platform you intend to support.

Response: We now emphasize that Linux is our intended platform of support, though general installation instructions are also provided that have been tested on both Linux and MacOS.

> Documentation
>
> I like a lot about the example configuration file (config_template/template_neat_config.yml) having instruction baked in, and this is almost the only thing that I need after installation to get things up and running.
>
> Suggestion 3: While not really hard to find, mentioning the existence of the template in README.md would be nice.

Response: This is now mentioned in `README.md`.

> Automatic tests are great and incorporated in GitHub Action, but some features could benefit from more coverage. For example, using the example data to run a minimal example caused a crash (ncsa/NEAT#161). Two pull requests (ncsa/NEAT#162 and ncsa/NEAT#163) were submitted and merged to address this issue.
>
> Suggestion 4: A higher-level integration test that take the example data for a quick run could be nice.

Response: We now add a higher-level integration test that uses a minimal example.

> In addition, ncsa/NEAT#162 got me into noticing that as a simulation tool, how error models and mutation models work appear not to have been documented beyond the original release manuscript, but given Table 1, I assume these have changed in the current version submitted for review. Therefore,
>
> Suggestion 5: How error models and mutation models work should be documented either in the manuscript or README.md.

Response: We now document how the error and mutation models work in `README.md`.

Addition: “Trinucleotides are identified in the reference genome and the variant file. The mutation model uses trinucleotide context and selects mutation sites and alternate alleles with a transition matrix. Frequencies of each trinucleotide transition are calculated and output as a pickle file. Mutations are simulated to reflect the same context-dependent biases as the training data. In NEAT 4.3.6, we have only made minor optimizations to improve the speed, and the underlying statistical models are similar to those described in the original NEAT manuscript.”

Addition: “The sequencing error model is stored as a pickled object, and during read generation, it introduces substitution, insertion, and deletion errors. Substitution alleles are sampled from a base-substitution transition matrix, while indel lengths are sampled from indel length distributions. The current `model-seq-err` utility estimates quality-score statistics and an overall average error rate from empirical FASTQ data (including FASTQ data derived from BAM files), and it packages these with the error model for simulation.”

> Suggestion 6: The error model and mutation models should be tested in the test suite. This would also help clarifying whether some behavior is intended. For example, the semantics of how_many appears to indicate that we are expecting a given number of alternative alleles to be present, but in neat/common/ploid_functions.py, the implementation seems to generate at most how_many alternative alleles since line 47 can select the same chromosome multiple times. I understand that whether a chromosome can be sampled with replacement is not right nor wrong, but it would be nice to have a clear intention so the users know what to expect if they wish to understand what was simulated with NEAT.

Response: We have included several tests of the error and mutational models in the test suite. The bugs and potential ambiguities corresponding to this have been corrected. We can confirm that we are sampling without replacement now. In terms of the question about ploidy, we note that the definition of “heterozygous” is ambiguous in ploidies above 2. We now indicate that for example, if ploidy is 25, “heterozygous” mutates 12 chromosomes (divide by two, and round down if necessary). Polyploidy was not handled in the original code. We expect most current users of NEAT to use ploidy = 2. Discussion of this ploidy-related ambiguity was added to `README.md`.

> Others
>
> Suggestion 7: A discussion of tools for the same or similar purposes is currently lacking. It could be nice to include the tools discussed in Alosaimi et al. and a functionality matrix to show case the feature set of these tools.

Response: Thank you for the feedback! We agree that a functionality matrix that showcases the feature set of different short-read simulators would be incredibly useful. Upon research into different tools, we discovered a paper entitled “Performance evaluation of six popular short-read simulators” by Milhaven & Pfeifer (2023) that creates functionality matrices in both of their key tables. We have elected to highlight Milhaven & Pfeifer’s findings about NEAT in comparison to similar tools to make use of neutral benchmarking evidence. We also discussed shortcomings of NEAT that the authors mention.

----------------------------------------------------------------------------------------------------

> @erik-whiting
>
> My review notes for the paper part only:
>
> Overall: Paper talks about version 4.2 but the repo is on 4.3 it looks like. Would it be easy to update the paper to talk about 4.3 or do we only want to introduce 4.2 with this paper?

Response: We had rolled out version 4.3 after submission of the manuscript and have updated the manuscript to match the repo.

> Summary: The summary doesn’t introduce the software at all, just gives an overview of some computational challenges around working with NGS data. Suggestion: Please at least briefly introduce NEAT here/

Response: We have addressed this point of feedback by briefly introducing NEAT in the summary.

Addition: “The NExt-generation sequencing Analysis Toolkit (NEAT) is an open-source Python package that creates such simulated sequencing datasets and was originally released as NEAT 2.0 in 2016. Here, we describe ongoing updates released collectively as NEAT 4.3, including new features and increased speed, accuracy, and usability.”

> Statement of Need: The statement of need section is more like a description of the changes made in this version of NEAT. You don’t talk about why this software suite is needed or how it compares to similar tools in this space.

Response: With multiple paragraphs, we now discuss why NEAT in particular is needed and is useful in comparison to similar tools. A discussion of similar tools in the field is now also provided in the “State of the field” section.

> State of the Field: Missing

Response: “State of the field” is now included in the paper as a separate section consisting of multiple paragraphs.

> Quality of writing is great and the references seem to be in order
>
> Here is the rest of my review:
>
> Functionality:
>
> Installation: I had to change the environment.yml file to not specify the libgcc version for my conda command to work correctly.
>
> Functionality: Good

Response: We have updated the installation instructions appropriately.

> Documentation:
>
> Statement of Need: I don’t see this anywhere?

Response: A version of the statement of need is now added to `README.md` in addition to the paper.

> Installation instructions: They mostly worked but I had to update the environment.yml file to not specify a libgcc version.
>
> Further, I think the installation instructions could be improved by including the git clone and cd commands. I.e.:
> > git clone git@github.com:ncsa/NEAT.git
> > cd NEAT
> > conda env create -f environment.yml -n neat
> … etc …

Response: We have updated the installation instructions appropriately.

> Automated tests: I see test scripts but I don’t see any hint of automation nor any documentation for how to run them. Is there a CI pipeline the public doesn’t see?

Response: The test scripts are now automated in addition to those in `.github/workflows/python-app.yml` when pushing or pulling between particular branches (currently set to `develop`, but we plan to update this in the near future), and there is now documentation at `CONTRIBUTING.md`.

> If I didn't mention something it's because it looks good and I checked it off in my checklist

----------------------------------------------------------------------------------------------------
