### 2026-03-05 31a5080
Version without the extra formatting now that the editorialbot method worked.
