### 2026-03-20 #9056 @bricoletc
@csoneson @keshav-gandhi I've gone through the revised submission and the authors have addressed all of my comments.
I'm happy with the revision and would like to provide few additional comments for the revised submission:

- Structure: for me the revised paper has all the right information but possibly not exactly in the right order, and in any case with missing sections (see [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)):
    - I would break out of `Statement of need` section after line 43, and start `Software Design` section here. You have a section called `Tables`, remove that header
    - For your two Tables, the main text points to them using numbers. Start your paragraphs stating which table you're discussing, and perhaps you could use bullet points, one per line (i.e. number) in the table you're discussing - but up to you for the bullet points.
    - State of the field section is very nice! From line 127, you should break it off though, and start the section `Research impact statement` as here you're discussing the (widespread) use of NEAT
- Line 45, remove 'since 2020' as in the summary you state the previous version of NEAT is from 2016, I found that 2020 mention confusing
- Line 76, 'variant insertion', are variants really 'integrated into simulated genomes', or just in the reads and VCF? If the latter, state that, as I find that sentence currently confusing
- Finally, I wasn't convinced by lines 61-63 regarding GC bias. So, you removed GC bias modelling module from NEAT, stating this was to streamline the codebase in the paper, and that it had 'minimal impact of on any final outputs in NEAT 2.0' in your review response above. Do you have a quantification of this minimal impact? I think it's fair to remove the GC bias module; [this paper](https://academic.oup.com/nar/article/40/10/e72/2411059) you cite states GC bias is more of a PCR, not a Illumina sequencing issue. But lines 61-63 suggest users have an alternative, by specifying which BED regions to sample from and not sample from; but this is binary, right? So you couldn't under-sample GC rich region this way? In this case I find that misleading and would state this feature somewhere else. If this BED sampling feature is not new in NEAT 4.3, I'd remove this mention completely.

Congrats and all the best!
