### 2026-03-21 #9056 @csoneson
Thanks @bricoletc and @erik-whiting! Could I ask you to please also tick off the remaining items in your checklists above, for completeness?

Regarding the missing sections in the paper mentioned by @bricoletc - note that JOSS recently updated its scope, but that this paper is being reviewed under the previous guidelines, as the review was initialized before the change. See [this post](https://github.com/openjournals/joss-reviews/issues/9056#issuecomment-3719110648) above for more details and links to the guidelines.
