### 2026-03-21 #9056 @chenyenchung
Hi all,

Thank you for the patience. All my previous suggestions and questions are covered in the current version, but I am with @bricoletc about the removal of GC bias control. Specifically, in the original implementation (tag: 2.1), a global GC-ratio-sampling rate relationship is modeled and applied to sampling. On the other hand, the current BED masking is binary so claiming "GC bias computation models the tendency of sequencing instruments to produce uneven coverage as a function of a local genomic region’s GC content, often underrepresenting regions of particularly high or low GC levels (Benjamini & Speed, 2012; Ross et al., 2013). To streamline the codebase, NEAT 4.3 removes the GC-bias module. _Instead, NEAT allows users to control coverage patterns via Browser Extensible Data (BED) region targeting._" (Line 57 - 62) is a bit misleading.

Saying so, I do agree that it can be argued that some newer sequencing platform has a much weaker sampling bias within moderate range of GC ratio (e.g., [Kim and colleagues](https://pmc.ncbi.nlm.nih.gov/articles/PMC7953489/) reported little bias from 20% to 80% for Illumina NovaSeq), but if this is the rationale, I believe the point would be clearer if made explicitly to back the "...minimal impact on the output..." claim.

Congratulations on the great work!

Best,
Yen
