### 2026-03-27 #9056 @keshav-gandhi
Hello @csoneson, @bricoletc, @chenyenchung, and @erik-whiting!

Thank you for your time, energy, and feedback in this second round of revisions.

Below each quote are our responses to your feedback:

----------------------------------------------------------------------------------------------------

@bricoletc

> Structure: for me the revised paper has all the right information but possibly not exactly in the right order, and in any case with missing sections (see [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)):

Response: We have updated the organization of our paper with the suggested changes.

- > I would break out of `Statement of need` section after line 43, and start `Software Design` section here. You have a section called `Tables`, remove that header

Response: A section entitled “`Software design`” has now been added after line 43 in place of the `Tables` section.

- > For your two Tables, the main text points to them using numbers. Start your paragraphs stating which table you're discussing, and perhaps you could use bullet points, one per line (i.e. number) in the table you're discussing - but up to you for the bullet points.

Response: To promote clarity, the text describing the changes outlined in **Table 1** has been organized with bullet points. **Table 2** was kept intact due to its comparative brevity. Sentences that introduce the changes in **Table 1** and **Table 2** have been added to the beginning of each respective sub-section.

- > State of the field section is very nice! From line 127, you should break it off though, and start the section Research impact statement as here you're discussing the (widespread) use of NEAT

Response: Thank you for the feedback! We have added the section entitled “`Research impact statement`” after this line.

> Line 45, remove 'since 2020' as in the summary you state the previous version of NEAT is from 2016, I found that 2020 mention confusing

Response: We have removed “since 2020” from this line.

> Line 76, 'variant insertion', are variants really 'integrated into simulated genomes', or just in the reads and VCF? If the latter, state that, as I find that sentence currently confusing

Response: We have addressed this feedback by clarifying that variants are incorporated into the reads and VCF.

Addition: “Variant insertion (7) handles how user-provided variants are incorporated into simulated reads and reported in the final VCF output. NEAT allows users to input their own VCF file of variants to include in the simulated reads. In the final simulated VCF, NEAT 4.3 preserves per-sample genotypes and allele-frequency information.”

> Finally, I wasn't convinced by lines 61-63 regarding GC bias. So, you removed GC bias modelling module from NEAT, stating this was to streamline the codebase in the paper, and that it had 'minimal impact of on any final outputs in NEAT 2.0' in your review response above. Do you have a quantification of this minimal impact? I think it's fair to remove the GC bias module; [this paper](https://academic.oup.com/nar/article/40/10/e72/2411059) you cite states GC bias is more of a PCR, not a Illumina sequencing issue. But lines 61-63 suggest users have an alternative, by specifying which BED regions to sample from and not sample from; but this is binary, right? So you couldn't under-sample GC rich region this way? In this case I find that misleading and would state this feature somewhere else. If this BED sampling feature is not new in NEAT 4.3, I'd remove this mention completely.

Response: Thank you for the feedback! We removed the potentially misleading statements regarding BED targeting. We do not currently have formal benchmarks quantifying the impact of GC bias on NEAT 2.0. Instead, we now describe the change with a more narrow claim: as suggested, we cite literature indicating that much of GC bias is an artifact of library preparation and amplification steps rather than the sequencing process itself, with relatively uniform coverage across the moderate GC range (20% to 60%) in a recent seven-platform benchmark.

Addition: “GC bias (2) is the tendency of sequencing data to have uneven coverage as a function of a local genomic region’s GC content, often underrepresenting regions of particularly high or low GC levels [@Benjamini:2012; @Ross:2013]. NEAT 4.3 removes NEAT 2.0’s module that models GC-dependent coverage bias. Evidence suggests that much of GC bias arises from library preparation and amplification steps rather than sequencing itself [@Benjamini:2012]. Additionally, recent benchmarks across seven sequencing platforms reported relatively uniform coverage across the moderate GC range (20% to 60%), which comprises about 95% of the human genome [@Kim:2021].”

----------------------------------------------------------------------------------------------------

@chenyenchung

> Thank you for the patience. All my previous suggestions and questions are covered in the current version, but I am with @bricoletc about the removal of GC bias control. Specifically, in the original implementation (tag: 2.1), a global GC-ratio-sampling rate relationship is modeled and applied to sampling. On the other hand, the current BED masking is binary so claiming "GC bias computation models the tendency of sequencing instruments to produce uneven coverage as a function of a local genomic region’s GC content, often underrepresenting regions of particularly high or low GC levels (Benjamini & Speed, 2012; Ross et al., 2013). To streamline the codebase, NEAT 4.3 removes the GC-bias module. _Instead, NEAT allows users to control coverage patterns via Browser Extensible Data (BED) region targeting._" (Line 57 - 62) is a bit misleading.
>
> Saying so, I do agree that it can be argued that some newer sequencing platform has a much weaker sampling bias within moderate range of GC ratio (e.g., [Kim and colleagues](https://pmc.ncbi.nlm.nih.gov/articles/PMC7953489/) reported little bias from 20% to 80% for Illumina NovaSeq), but if this is the rationale, I believe the point would be clearer if made explicitly to back the "...minimal impact on the output..." claim.

Response: Thank you for the feedback! We have removed all mentions of the binary BED masking from the main text and **Table 1**. We now frame this section that describes the module’s removal in light of advances in short-read sequencing platforms that appear to have reduced GC-associated coverage bias across moderate GC ranges, as shown by Kim et al. (2021).
