### 2026-04-10 #9056 @chenyenchung
> Response: Thank you for the feedback! We have removed all mentions of the binary BED masking from the main text and Table 1. We now frame this section that describes the module’s removal in light of advances in short-read sequencing platforms that appear to have reduced GC-associated coverage bias across moderate GC ranges, as shown by Kim et al. (2021).

Looks good to me. I have no further questions. Kudos on the great work!
