### 2026-04-11 #9056 @csoneson
@keshav-gandhi - the paper is currently much longer than the JOSS guidelines (250-1000 words, see [here](https://joss.readthedocs.io/en/pre-2026-submission/paper.html#what-should-my-paper-contain)). Could you please have a go at shortening it to be more compliant with the guidelines (you can generate a new proof with `@editorialbot generate pdf`)? Thanks!
