### 2026-04-19 #9056 @csoneson
Thanks @keshav-gandhi - I have gone through the submission and overall this looks very good to me! The only thing I found is that some references need case protection (addition of `{}` to avoid capitals being rendered in lower-case). I found this in:
* Ahmed et al (2019)
* Caboche et al (2014)
* Gourlé et al (2019)
* Hu et al (2012)
* Jandrasits et al (2019)
* Lefouili et al (2022)
* Rhie et al (2023)
* Schmeing et al (2021)

Could you fix this and then generate a new proof with `@editorialbot generate pdf`? I will also make another post below outlining the next steps after that.
