### 2026-05-03 #9056 @keshav-gandhi
Hi @csoneson!

We have completed the following tasks:

> - Double check authors and affiliations (including ORCIDs)

Affiliations and ORCIDs have been double-checked and updated accordingly.

> - Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

NEAT v4.4 has just been released, and the paper was updated. Future enumerations of NEAT version will follow the pattern v4.4.X.

> - Archive the release on Zenodo/figshare/etc and post the DOI here.

Here is our Zenodo DOI: 10.5281/zenodo.19662441

> - Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

We confirm that the title, author list, and ORCIDs match between the most recent version of the JOSS paper and Zenodo archive.

> - Make sure that the license listed for the archive is the same as the software license.

We confirm that the licenses are the same (BSD 3-Clause License).
