### 2026-05-03 #9056 @csoneson
Thanks @keshav-gandhi. Could you please also make sure that the version tag is the same for the GitHub release (currently `4.4`) and on Zenodo (currently `v4.4`)?
