### 2026-05-03 #9056 @csoneson
@openjournals/dev Could you help me figure out the reason why the JATS validation fails here? 🙏🏻
