### 2026-05-04 #9056 @arfon
It looks to be an issue with the table formatting. To start, I would suggest trying removing the # on the first line of the table and doing `recommend-accept` again
